% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{addCtryPolyIdx}
\alias{addCtryPolyIdx}
\title{Add an index column to all layers of a polygon}
\usage{
addCtryPolyIdx(
  ctryCode,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"),
  custPolyPath = NULL
)
}
\arguments{
\item{ctryCode}{The ISO3 ctryCode of the country polygon to download}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
None
}
\description{
Add an index column to all layers of a polygon
}
\examples{
\dontrun{
Rnightlights:::addCtryPolyIdx(ctryCode="KEN")
}

}
