% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{allValid}
\alias{allValid}
\title{Check if a vector/list of values given is valid as per the given validation function}
\usage{
allValid(testData, testFun)
}
\arguments{
\item{testData}{the list/vector of values to validate}

\item{testFun}{the validation function to test each value of testData against}
}
\value{
TRUE/FALSE
}
\description{
Check if a vector/list of values given is valid as per the given validation function. 
    The function will also print a warning showing the values that are invalid. One can 
    stop the warning being printed by wrapping the function in the \code{suppressWarnings} 
    function.
}
\examples{

ctryCodes <- c("KE", "UGA", "RWA", "TZ")
allValid(ctryCodes, validCtryCode)
 returns TRUE

}

