% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{allValidCtryAdmLvls}
\alias{allValidCtryAdmLvls}
\title{Checks if all ctry admLevels are valid}
\usage{
allValidCtryAdmLvls(ctryCode, admLevels)
}
\arguments{
\item{ctryCode}{\code{character} The ctryCode of the country of interest}

\item{admLevels}{\code{character} The admLevel(s) to search for}
}
\value{
\code{logical} whether inputted admLevels are valid
}
\description{
Checks if all ctry admLevels are valid
}
\examples{
\dontrun{
validCtryAdmLvls("KEN", "adm0")
#returns "KEN_adm1"
}

}
