% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrycode.R
\name{allValidCtryCodes}
\alias{allValidCtryCodes}
\title{Check if all ctryCodes are valid}
\usage{
allValidCtryCodes(ctryCodes)
}
\arguments{
\item{ctryCodes}{the ISO3 country codes to validate}
}
\value{
TRUE/FALSE
}
\description{
Check if all ctryCodes are valid
}
\examples{
Rnightlights:::allValidCtryCodes(c("BDI", "ETH", "KEN", "RWA", "TZA", "UGA")) #returns TRUE

#returns FALSE. "United Arab Emirates" ISO3 code = "ARE"
Rnightlights:::allValidCtryCodes(c("UGA", "UAE"))

}
