% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{allValidNlPeriods}
\alias{allValidNlPeriods}
\title{Check if all nlPeriods are valid for given nlTypes}
\usage{
allValidNlPeriods(nlPeriods, nlTypes)
}
\arguments{
\item{nlPeriods}{\code{vector or list of character vectors} The nlPeriods of interest}

\item{nlTypes}{\code{vector or list of character vectors} type of nightlight}
}
\value{
\code{logical} TRUE/FALSE
}
\description{
Check if all nlPeriods are valid for given nlTypes
}
\examples{
Rnightlights:::allValidNlPeriods(c("201401", "201402"),"VIIRS.M")
 #returns TRUE

Rnightlights:::allValidNlPeriods("201203","VIIRS.M")
 #returns FALSE

Rnightlights:::allValidNlPeriods("2012","OLS.Y")
 #returns TRUE

}
