% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nltype.R
\name{allValidNlTypes}
\alias{allValidNlTypes}
\title{Checks if all given character strings are valid nlTypes}
\usage{
allValidNlTypes(nlTypes)
}
\arguments{
\item{nlTypes}{\code{character vector} string The nlTypes}
}
\value{
\code{logical vector} whether the strings are valid nlTypes
}
\description{
Checks if all given character strings are valid nlTypes
}
\examples{
Rnightlights:::allValidNlTypes("VIIRS.D") #returns TRUE

Rnightlights:::allValidNlTypes("VIIRS") #returns FALSE

}
