% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{cleanupNtLts}
\alias{cleanupNtLts}
\title{Clean up the environment after processing (Not yet implemented)}
\usage{
cleanupNtLts()
}
\description{
Clean up the environment after processing (Not yet implemented)
}
\examples{
 cleanup()

}
