% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{createCtryStruct}
\alias{createCtryStruct}
\title{Save ctry admin structure to text file for faster access}
\usage{
createCtryStruct(ctryCode)
}
\arguments{
\item{ctryCode}{The ISO3 ctryCode of the country}
}
\value{
None
}
\description{
Save ctry admin structure to text file for faster access
}
\examples{
\dontrun{
  Rnightlights:::createCtryStruct("KEN")
}

}
