% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapath.R
\name{createNlDataDirs}
\alias{createNlDataDirs}
\title{Create required data subdirectories in the root data path}
\usage{
createNlDataDirs()
}
\value{
None
}
\description{
Create required data subdirectories in the root data path
}
\examples{
  \dontrun{
  Rnightlights:::createNlDataDirs()
  }
    
}
