% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasflares.R
\name{createNlGasFlares}
\alias{createNlGasFlares}
\title{Get the path to the saved world gasflare RDS}
\usage{
createNlGasFlares()
}
\value{
None
}
\description{
Get the path to the saved world gasflare RDS
}
\examples{
\dontrun{
  getNlGasFlaresRdsFnamePath(ctryCode="NGA")
#returns path to the world RDS
}

}
