% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{createNlTilesSpPolysDF}
\alias{createNlTilesSpPolysDF}
\title{Creates a tile Spatial Polygons DataFrame from the \code{"nlTiles"} dataframe}
\usage{
createNlTilesSpPolysDF()
}
\value{
TRUE/FALSE
}
\description{
Creates a Spatial Polygons DataFrame from the \code{"nlTiles"} dataframe of VIIRS tiles
}
\examples{
  tilesSpPolysDFs <- Rnightlights:::createNlTilesSpPolysDF()

}
