% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{ctryCodeToName}
\alias{ctryCodeToName}
\title{Convert a country ISO3 code to the full name}
\usage{
ctryCodeToName(ctryCode)
}
\arguments{
\item{ctryCode}{}
}
\value{
Character full country name
}
\description{
Convert a country ISO3 code to the full name. Exposes the rworldmap function 
    isoToName(ctryCode). #rworldmap::isoToName can resolve 2-letter ctryCodes 
    but we only want 3-letter ISO3 codes
}
\examples{
ctryCodeToName("KEN")

}
