% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrycode.R
\name{ctryNameToCode}
\alias{ctryNameToCode}
\title{Convert a country name to its ISO3 code}
\usage{
ctryNameToCode(ctryNames)
}
\arguments{
\item{ctryName}{Chracter string common name of a country}
}
\value{
Character ISO3 ctryCode if found else NA
}
\description{
Convert a country name to its ISO3 code. Searches the rworldmap map data.
    With no parameters returns a list of ctryNames and their corresponding
    codes as given by rworldMap
}
\examples{
ctryNameToCode("kenya") #returns "KEN"

ctryNameToCode("ken") #returns "KEN"
  
ctryNameToCode("jamaica") #returns JAM

}
