% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{ctryPolyAdmColNames}
\alias{ctryPolyAdmColNames}
\title{Check if a month number is valid for a given nightlight type}
\usage{
ctryPolyAdmColNames(ctryPolyAdmLevels, nLyrs)
}
\arguments{
\item{monthNum}{the month in "MM" format e.g. Jan="01", Feb="02"}

\item{nlType}{type of nightlight either "VIIRS" or "OLS"}
}
\value{
TRUE/FALSE
}
\description{
Check if a month number is valid for a given nightlight type. Note month num is only valid for
"VIIRS" nightlight type
}
\examples{
validNlMonthNum("01","VIIRS")
 returns TRUE

validNlMonthNum("13","VIIRS")
 returns FALSE

validNlMonthNum("01","OLS")
 returns FALSE

}
