% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{dateToNlPeriod}
\alias{dateToNlPeriod}
\title{Convert date to nlPeriod}
\usage{
dateToNlPeriod(dt, nlType)
}
\arguments{
\item{dt}{the date to convert}

\item{nlType}{the nlType to use}
}
\value{
an nlPeriod vector
}
\description{
Convert date to nlPeriod
}
\examples{
dateToNlPeriod(dt = "2012-04-01", nlType = "VIIRS.M")
#returns "201204"

}
