% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{deleteSavedNlStat}
\alias{deleteSavedNlStat}
\title{Delete a previously saved nlStat}
\usage{
deleteSavedNlStat(nlStat)
}
\arguments{
\item{nlStat}{The nlStat to delete}
}
\value{
an nlPeriod vector
}
\description{
Delete a previously saved nlStat
}
\examples{
deleteSavedNlStat(nlStat)
#returns "201204"

}
