% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{dnldCtryPoly}
\alias{dnldCtryPoly}
\title{Download a country's polygon shapefile from \url{http://gadm.org}}
\usage{
dnldCtryPoly(ctryCode)
}
\arguments{
\item{ctryCode}{The ISO3 ctryCode of the country polygon to download}
}
\value{
TRUE/FALSE Success/Failure of the download
}
\description{
Download a country's polygon shapefile from \url{http://gadm.org}
}
\examples{
\dontrun{
Rnightlights:::dnldCtryPoly("KEN")
}

}
