% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{dnldGADMCtrySpRds}
\alias{dnldGADMCtrySpRds}
\title{Download a country's polygon RDS files from \url{http://gadm.org}}
\usage{
dnldGADMCtrySpRds(
  ctryCode,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"),
  downloadMethod = pkgOptions("downloadMethod")
)
}
\arguments{
\item{ctryCode}{The ISO3 ctryCode of the country polygon to download}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{downloadMethod}{The method used to download polygons}
}
\value{
TRUE/FALSE Success/Failure of the download
}
\description{
Download a country's polygon RDS files from \url{http://gadm.org} and
    combine them into one RDS to match other polygon downloads
}
\examples{
\dontrun{
Rnightlights:::dnldCtryPoly("KEN")
}

}
