% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.Last.lib}
\alias{.Last.lib}
\title{Clean up the environment}
\usage{
.Last.lib(libpath)
}
\description{
Clean up the environment and delete temp files when the package is
    detached.
    
    Use .Last.lib rather than .onDetach or .onUnload which may not be run
    at the end of the session. And rather than reg.finalizer which is
    called when gc() runs which we do not want to happen for nlCleanup()
    especially since we do call gc() in some instances.
}
