% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Set global variables in .onAttach since we want the package to be
    usable without active loading via library(Rnightlights)}
\usage{
.onAttach(libname, pkgname)
}
\description{
Set global variables in .onAttach since we want the package to be
    usable without active loading via library(Rnightlights)
}
