% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.RnightlightsEnv}
\alias{.RnightlightsEnv}
\title{A package-wide hidden environment}
\format{An object of class \code{environment} of length 0.}
\usage{
.RnightlightsEnv
}
\description{
A package-wide hidden environment used to store variables/settings and
    functions to retrieve them
}
\keyword{datasets}
