% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadnltiles.R
\name{downloadNlTilesOLS}
\alias{downloadNlTilesOLS}
\title{Download OLS nightlight tile}
\usage{
downloadNlTilesOLS(
  nlPeriod,
  downloadMethod = pkgOptions("downloadMethod"),
  nlType = "OLS.Y",
  configName = pkgOptions(paste0("configName_", nlType)),
  extension,
  multiTileStrategy = pkgOptions("multiTileStrategy")
)
}
\arguments{
\item{nlPeriod}{the nlPeriod of the tile}

\item{downloadMethod}{The method to use for download.}

\item{nlType}{A character string of nlType}

\item{configName}{character the config shortname of the raster to process}

\item{extension}{character the extension of the raster to process}

\item{multiTileStrategy}{character How to handle multiple tiles per nlPeriod}
}
\value{
TRUE/FALSE Whether the download was successful
}
\description{
Download OLS nightlight tile
}
\examples{
\dontrun{
if(Rnightlights:::downloadNlTilesOLS("201405"))
  print("download successful")
  }

}
