% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{equalNlStats}
\alias{equalNlStats}
\title{Check if two nlStats are equal}
\usage{
equalNlStats(nlStatName1, nlStatName2)
}
\arguments{
\item{nlStatName1}{The name of the first nlStat function to compare}

\item{nlStatName2}{The name of the second nlStat function to compare}
}
\value{
logical if the two nStats are equal
}
\description{
Check if two nlStats are equal
}
\examples{
fn1 <- function(x) sum(x)
fn2 <- function(x) sum(x, na.rm=T)
Rnightlights:::equalNlStats("fn1", "fn2")
#returns FALSE

}
