% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{existsCtryCodeTiles}
\alias{existsCtryCodeTiles}
\title{Check if the variable ctryCodeTiles exists}
\usage{
existsCtryCodeTiles()
}
\value{
TRUE/FALSE
}
\description{
Check if the variable ctryCodeTiles, which is a mapping between countryCodes and the tiles it intersects with, exists as created by \code{mapCtryPolyToTiles()}.
}
\examples{
if(!existsCtryCodeTiles())
  mapCtryPolyToTiles()

}

