% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrynldata.R
\name{existsCtryNlData1}
\alias{existsCtryNlData1}
\title{Check if VIIRS nightlight stats exist locally}
\usage{
existsCtryNlData1(ctryCode, admLevel, nlPeriod, stat, nlType)
}
\arguments{
\item{ctryCode}{character the ISO3 code of the country}

\item{nlPeriod}{character the nlPeriod}

\item{stat}{character the stat to check for}

\item{nlType}{character the nlType}
}
\value{
TRUE/FALSE
}
\description{
Check if VIIRS nightlight data for the country exists in the country 
    nightlight data file. First checks if the country nightlight data
    file exists.
}
\examples{
Rnightlights:::existsCtryNlData("KEN", "201401", "sum", "VIIRS.M")

}
