% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{existsCtryNlDataFile}
\alias{existsCtryNlDataFile}
\title{Check if a country's data file exists}
\usage{
existsCtryNlDataFile(ctryCode)
}
\arguments{
\item{ctryCode}{the ISO3 country code}
}
\value{
TRUE/FALSE
}
\description{
Check if a country's data file exists
}
\examples{
ctryCode <- "KEN"
if(existsCtryNlDataFile(ctryCode))
 message("Data file for ", ctryCode, " found")

}

