% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{existsCtryNlDataVIIRS}
\alias{existsCtryNlDataVIIRS}
\title{Check if VIIRS nightlight data for a country in a given year and month already exists locally}
\usage{
existsCtryNlDataVIIRS(ctryCode, nlYearMonth, stat)
}
\arguments{
\item{ctryCode}{character the ISO3 code of the country}

\item{nlYearMonth}{character the year and month concatenated}

\item{stat}{character the stat to check for}
}
\value{
TRUE/FALSE
}
\description{
Check if VIIRS nightlight data for the country exists in the country nightlight data file. 
    First checks if the country nightlight data file exists.
}
\examples{
existsCtryNlDataVIIRS("KEN", "201204")

}

