% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{existsCtryPoly}
\alias{existsCtryPoly}
\title{Checks if a country polygon exists}
\usage{
existsCtryPoly(ctryCode, gadmVersion = pkgOptions("gadmVersion"))
}
\arguments{
\item{ctryCode}{\code{character} The ctryCode of the country of interest}
}
\description{
Checks if a country polygon exists
}
\examples{
\dontrun{
existCtryPoly("KEN")
#returns TRUE/FALSE
}

}
