% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasflares.R
\name{existsNlCtryGasFlaresRds}
\alias{existsNlCtryGasFlaresRds}
\title{Checks if the country gasflares rds exists}
\usage{
existsNlCtryGasFlaresRds(ctryCode)
}
\arguments{
\item{ctryCode}{\code{character} The ctryCode of the country of interest}
}
\value{
\code{logical} whether the country has gasflares
}
\description{
Checks if the country gasflares rds exists
}
\examples{
\dontrun{
  existsNlCtryGasFlaresRds(ctryCode="NGA")
  #returns TRUE
}

}
