% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasflares.R
\name{existsNlGasFlaresRds}
\alias{existsNlGasFlaresRds}
\title{Check if the world gasflare RDS exists}
\usage{
existsNlGasFlaresRds()
}
\value{
\code{logical} if the gasflare rds exists
}
\description{
Check if the world gasflare RDS exists
}
\examples{
\dontrun{
  existsNlGasFlaresRds()
#returns TRUE/FALSE
}

}
