% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{existsNlTiles}
\alias{existsNlTiles}
\title{Checks if the nlTiles data.frame variable exists}
\usage{
existsNlTiles()
}
\value{
TRUE/FALSE
}
\description{
Checks if the nlTiles data.frame variable exists in the environment and that it is not null or empty. This is used by other functions that depend on the nlTiles data.frame to check if it exists. If it doesn't they may create the data.frame.
}
\examples{
if(!existsNlTiles())
  nlTiles <- getNlTiles()

}
