% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{existsPolyFnamePath}
\alias{existsPolyFnamePath}
\title{Check if the decompressed country polygon has been downloaded and stored in the polygon folder}
\usage{
existsPolyFnamePath(ctryCode)
}
\arguments{
\item{ctryCode}{}
}
\value{
TRUE/FALSE
}
\description{
Check if the decompressed country polygon has been downloaded and stored in the polygon folder
}
\examples{
existsPolyFnamePath("KEN")
 TRUE/FALSE

}
