% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{existsPolyFnameZip}
\alias{existsPolyFnameZip}
\title{Check if the compressed country polygon has been downloaded and stored in the polygon folder}
\usage{
existsPolyFnameZip(ctryCode)
}
\arguments{
\item{ctryCode}{The ctryCode of interest}
}
\value{
TRUE/FALSE
}
\description{
Check if the compressed country polygon has been downloaded and stored in the polygon folder
}
\examples{
Rnightlights:::existsPolyFnameZip("KEN")
 #returns TRUE/FALSE

}
