% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{existsSavedNlStat}
\alias{existsSavedNlStat}
\title{Check whether an nlStat exists in the saved nlStats}
\usage{
existsSavedNlStat(nlStatSig, nlStatHash)
}
\arguments{
\item{nlStatHash}{character The hash of the nlStat to check}

\item{nlStatName}{character The signature of the nlStat to check}
}
\value{
logical Whether the nlStat was found in the saved nlStats
}
\description{
Check whether an nlStat exists in the saved nlStats given the nlStat
    signature and the nlStatHash to ensure a unique hit
}
\examples{
\dontrun{
  existsSavedNlStat(nlStatName = "sum()", nlStatHash = "f0fbe35d81578311ba8f362137832e779b7b4f39")
  #returns TRUE/FALSE
}

}
