% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{existsSavedNlStatHash}
\alias{existsSavedNlStatHash}
\title{Check whether an nlStat exists in the saved nlStats given the nlStatHash}
\usage{
existsSavedNlStatHash(nlStatHash)
}
\arguments{
\item{nlStatHash}{character The hash of the nlStat to check}
}
\value{
logical Whether the nlStatHash was found in the saved nlStats
}
\description{
Check whether an nlStat exists in the saved nlStats given the nlStatHash
}
\examples{
\dontrun{
  existsSavedNlStatHash(nlStatHash = "f0fbe35d81578311ba8f362137832e779b7b4f39")
  #returns TRUE/FALSE
}

}
