% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{existsSavedNlStatName}
\alias{existsSavedNlStatName}
\title{Check whether an nlStat exists in the saved nlStats}
\usage{
existsSavedNlStatName(nlStatName)
}
\arguments{
\item{nlStatName}{character The name of the nlStat to check}
}
\value{
logical Whether the nlStat was found in the saved nlStats
}
\description{
Check whether an nlStat exists in the saved nlStats
}
\examples{
\dontrun{
  existsSavedNlStatName(nlStatName = "sum")
  #returns TRUE/FALSE
}

}
