% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{existsSavedNlStatSig}
\alias{existsSavedNlStatSig}
\title{Check whether an nlStat exists in the saved nlStats}
\usage{
existsSavedNlStatSig(nlStatSig)
}
\arguments{
\item{nlStatSig}{character The name of the nlStat to check}
}
\value{
logical Whether the nlStat was found in the saved nlStats
}
\description{
Check whether an nlStat exists in the saved nlStats given the signature
}
\examples{
\dontrun{
  existsSavedNlStatSig(nlStatSig = "sum()")
  #returns TRUE/FALSE
}

}
