% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{existsTilesSpPolysDFs}
\alias{existsTilesSpPolysDFs}
\title{Checks if the variable \code{"tilesSpPolysDFs"} exists}
\usage{
existsTilesSpPolysDFs()
}
\value{
TRUE/FALSE
}
\description{
Checks if the \code{"tilesSpPolysDFs"} (tile Spatial Polygons DataFrame) variable exists in the environment   and that it is not null. This is used by other functions that depend on the \code{"tilesSpPolysDFs"} data.frame to check if it exists. If it doesn't they can create the data.frame.
}
\examples{
if(!existsTilesSpPolysDFs())
  tilesSpPolysDFs <- createNlTilesSpPolysDF()

}

