% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fnAggRadGdal}
\alias{fnAggRadGdal}
\title{Calculate zonal statistics using GDAL. Faster than fnAggRadRast for large polygons.}
\usage{
fnAggRadGdal(ctryCode, ctryPoly, nlPeriod, fnStats = pkgOptions("stats"),
  nlType)
}
\arguments{
\item{ctryCode}{character string the ISO3 country code to be processed}

\item{ctryPoly}{Polygon the loaded country polygon layer}

\item{nlPeriod}{character string the nlPeriod to be processed}

\item{fnStats}{character vector The stats to calculate}

\item{nlType}{the nlType of interest}
}
\value{
data.frame of polygon attributes and the calculated stats
}
\description{
Calculate zonal statistics. Alternative to fnAggRadRast using GDAL. Faster for large polygons. 
    Modified from \url{http://www.guru-gis.net/efficient-zonal-statistics-using-r-and-gdal/}
}
\examples{
\dontrun{validNlMonthNum("01","VIIRS")}
 #returns TRUE

}
