% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{fnSumAvgRadGdal}
\alias{fnSumAvgRadGdal}
\title{Calculate zonal statistics using GDAL. Faster than fnSumAvgRadRast for large polygons.}
\usage{
fnSumAvgRadGdal(ctryCode, ctryPoly, nlYearMonth, fnStats = stats)
}
\arguments{
\item{ctryCode}{the ISO3 country code to be processed}

\item{ctryPoly}{the loaded country polygon layer}

\item{nlYearMonth}{the concatenated year+month to be processed}
}
\value{
data.frame of polygon attributes and the calculated stats
}
\description{
Calculate zonal statistics. Alternative to fnSumAvgRadRast using GDAL. Faster for large polygons. 
    Modified from \url{http://www.guru-gis.net/efficient-zonal-statistics-using-r-and-gdal/}
}
\examples{
validNlMonthNum("01","VIIRS")
 returns TRUE

}

