% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getAllGadmPolyTypes}
\alias{getAllGadmPolyTypes}
\title{Return a vector of GADM polyTypes}
\usage{
getAllGadmPolyTypes()
}
\value{
character vector valid GADM polyTypes
}
\description{
Return a vector of GADM polyTypes
}
\examples{

getAllGadmPolyTypes()

}
