% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getAllGadmVersions}
\alias{getAllGadmVersions}
\title{Return a vector of GADM versions}
\usage{
getAllGadmVersions()
}
\value{
character vector valid GADM versions
}
\description{
Return a vector of GADM versions
}
\examples{
\dontrun{
Rnightlights:::getAllGadmVersions()
}

}
