% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrycode.R
\name{getAllNlCtryCodes}
\alias{getAllNlCtryCodes}
\title{Check if a month number is valid for a given nightlight type}
\usage{
getAllNlCtryCodes(omit = "none")
}
\arguments{
\item{omit}{The ctryCodes to exclude from processing. Else keywords missing=ctryCodes
that are in rworldmap but not on gadm long=ctryCodes that take very long to process}
}
\value{
character vector of country codes
}
\description{
Check if a month number is valid for a given nightlight type. Note month num is only valid for
"VIIRS" nightlight type
}
