% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{getAllNlPeriods}
\alias{getAllNlPeriods}
\title{Generate a list of all possible nlPeriods for a given nlType}
\usage{
getAllNlPeriods(nlTypes)
}
\arguments{
\item{nlType}{type of nightlight either "VIIRS" or "OLS"}
}
\value{
character vector list of nlPeriods
}
\description{
Generate a list of all possible nlPeriods for a given nlType
}
\examples{
getAllNlPeriods("OLS.Y")
 #returns a vector of all years from 1994 to 2013

getAllNlPeriods("VIIRS.M")
 #returns a vector of all yearMonths from 201401 to present

}
