% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nltype.R
\name{getAllNlTypes}
\alias{getAllNlTypes}
\title{Lists supported nightlight types}
\usage{
getAllNlTypes()
}
\value{
\code{character} vector of supported nlTypes
}
\description{
Lists supported nightlight types. The names are in the form
    "nlType.avgPeriod" where avgPeriod is the period over which the
    raw data is averaged i.e. Daily (D), Monthly (M) and Yearly (Y).
    Currently the main nlTypes are "OLS" and "VIIRS". OLS only has yearly
    averaged data while VIIRS has daily, monthly and yearly
    e.g. VIIRS daily = "VIIRS.D", OLS yearly = "OLS.Y"
}
\examples{
getAllNlTypes()

}
