% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlurl.R
\name{getAllNlUrlOLS}
\alias{getAllNlUrlOLS}
\title{Function to return the url of the OLS tile to download}
\usage{
getAllNlUrlOLS(nlTypes, configNames = "", extensions)
}
\arguments{
\item{nlTypes}{The nlTypes for which to return the tile download URL}

\item{configNames}{character filter by configNames if specified}

\item{extensions}{character filter by extensions if specified}
}
\value{
character string Url of the OLS tile file
}
\description{
Function to return the url of the OLS tile to download given the year
}
\examples{
\dontrun{
  getAllNlUrlOLS()
}

}
