% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlurl.R
\name{getAllNlUrlVIIRS}
\alias{getAllNlUrlVIIRS}
\title{Function to return the url of the VIIRS tile to download}
\usage{
getAllNlUrlVIIRS(nlTypes, configNames)
}
\arguments{
\item{nlTypes}{character the nlTypes to consider}

\item{configNames}{character the configNames to consider}
}
\value{
Character string Url of the VIIRS tile file
}
\description{
Function to return the url of the VIIRS tile to download given the year, month, and nlTile index
}
\examples{
\dontrun{
tileUrl <- Rnightlights:::getAllNlUrlVIIRS("20171231", "1", "VIIRS.D")

tileUrl <- Rnightlights:::getAllNlUrlVIIRS("201401", "1", "VIIRS.M")

tileUrl <- Rnightlights:::getAllNlUrlVIIRS("2015", "1", "VIIRS.Y")
}

}
