% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getAllNlYearMonthsTiles}
\alias{getAllNlYearMonthsTiles}
\title{Downloads all the listed VIIRS tiles for a particular year, month}
\usage{
getAllNlYearMonthsTiles(nlYearMonths, tileList)
}
\arguments{
\item{nlYearMonths}{character vector a list of the nlYearMonths to process in the format "YYYYMM"}

\item{tileList}{character vector a list of tile numbers (1-6) to download}
}
\value{
TRUE/FALSE whether the download was successful
}
\description{
Downloads all the listed VIIRS tiles for a particular year, month
}
\examples{
#download tiles for "KEN" which are tiles 2 and 5 for the specified time periods
getAllNlYearMonthsTiles(c("201201", "201202", "201205"), c(2, 5))

#same as above but getting the tileList automatically
getAllNlYearMonthsTiles(c("201201", "201202", "201205"), tileName2Idx(getCtryCodeTileList("KEN")))
#returns TRUE if ALL downloads were successful

}

