% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBatchBytes}
\alias{getBatchBytes}
\title{Calculate the RAM to provide to the package for gdal calculations}
\usage{
getBatchBytes(freeRAM = pkgOptions("batchBytes"))
}
\arguments{
\item{freeRAM}{the amount of available RAM to consider}
}
\description{
Calculate the RAM to provide to the package to calculate when
    pkgOption("extractMethod") == "gdal"
}
