% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadnltiles.R
\name{getCredentialsEOG}
\alias{getCredentialsEOG}
\title{Retrieve credentials required for download from the EOG site}
\usage{
getCredentialsEOG(
  credFile = file.path(getNlDataPathFull(), pkgOptions("EOG_CredFile"))
)
}
\arguments{
\item{credFile}{(character) Path to a file containing the credentials
obtained from registration at 
\url{https://eogauth.mines.edu/auth/realms/master/account/}}
}
\description{
Retrieve credentials required for download from the new EOG site at 
    University of Colorado, Dept of Mines
}
