% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getCRS}
\alias{getCRS}
\title{Specify the wgs84 CRS globally}
\usage{
getCRS()
}
\description{
Specify the wgs84 CRS which is currently the only CRS we support
   for use throughout the package. Saves the proj4 wgs84 string to the
   .RnightlightsEnv package hidden environment
}
