% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfuns.R
\name{getCtryCodesWithData}
\alias{getCtryCodesWithData}
\title{Return ctryCodes that have data}
\usage{
getCtryCodesWithData()
}
\value{
character vector valid ctryCodes
}
\description{
Return ctryCodes that have data
}
\examples{
\dontrun{
 getCtryCodesWithData()
}

}
