% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getCtryGFMaskOutputFnamePath}
\alias{getCtryGFMaskOutputFnamePath}
\title{Get the full path to the file where the cropped VIIRS country mask is stored.}
\usage{
getCtryGFMaskOutputFnamePath(
  ctryCode,
  nlType,
  nlPeriod,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"),
  custPolyPath = NULL
)
}
\arguments{
\item{ctryCode}{(character) the ctryCode of interest}

\item{nlType}{(character) the nlType of interest}

\item{nlPeriod}{(character) the nlPeriod of interest}

\item{gadmVersion}{(character) The GADM version to use}

\item{gadmPolyType}{(character) The format of polygons to download from GADM}

\item{custPolyPath}{(character) The path to a custom polygon as an alternative
to using GADM polygons}
}
\value{
(character) full path to the cropped VIIRS country raster for a country and a given year and month
}
\description{
Get the full path to the file where the cropped VIIRS country mask is stored. This file is created
    when processing the country before extracting the data. It can be used to re-process a country much faster
}
\examples{
\dontrun{
getCtryRasterOutputFnamePath("KEN","VIIRS.M", "201412")
}

}
