% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getCtryNlDataColName}
\alias{getCtryNlDataColName}
\title{Construct the name of a nightlight data column given the nightlight type and yearMonth}
\usage{
getCtryNlDataColName(nlPeriod, stat, nlType)
}
\arguments{
\item{stat}{}

\item{nlType}{character vector. the type of nightlight i.e. "OLS" or VIIRS. Default=VIIRS}

\item{nlYearMonth}{character vector. the yearMonth (concat year + month) in the format 
"YYYYMM" e.g. "201203"}
}
\value{
character string
}
\description{
Construct the name of a nightlight data column given the nightlight type and yearMonth. Used in
    creating and retrieving data columns from the nightlight data file
}
\examples{
ctryCode <- "KEN"
dt <- read.csv(getCtryNlDataFnamePath(ctryCode))
dt <- dt[,getCtryNlDataColName("201612", "sum", nlType="VIIRS")]
  #returns the column "NL_201612_SUM" if it exists in the KEN data file
}
