% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getCtryNlDataFname}
\alias{getCtryNlDataFname}
\title{Check if a month number is valid for a given nightlight type}
\usage{
getCtryNlDataFname(ctryCode)
}
\arguments{
\item{ctryCode}{}
}
\value{
Character filename of the country data file
}
\description{
Get the name of the data file. This function can be altered to name the file as 
    required and consistently retrieve the name. Used in the function getCtryNlDataFnamePath 
    to concat the directory path and this filename. Currently all nlTypes are stored in one 
    file. Can be altered to separate VIIRS and OLS data files for example.
}
\examples{
ctryCode <- "KEN"
getCtryNlDataFname(ctryCode)
 #returns name of the ctry data file

}
