% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrynldata.R
\name{getCtryNlDataFnamePath}
\alias{getCtryNlDataFnamePath}
\title{Construct the full path to save the file containing the country data}
\usage{
getCtryNlDataFnamePath(ctryCode)
}
\arguments{
\item{ctryCode}{character string The ctryCode of interest}
}
\value{
Character string the full path to the data file of a country
}
\description{
Construct the full path to save the file containing the country data. Note
    it does not indicate if the file exists
}
\examples{
#get the full path to the file containing data for KEN
getCtryNlDataFnamePath("KEN")

Rnightlights:::getCtryNlDataFnamePath("KEN")
 #returns full path of the nightlight data file for the country

#@export only due to exploreData() shiny app
}
