% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfuns.R
\name{getCtryNlTypes}
\alias{getCtryNlTypes}
\title{Return the unique nlTypes available in the data}
\usage{
getCtryNlTypes(
  countries,
  admLevel,
  polySrc = "GADM",
  polyVer = "3.6",
  polyType = "SHPZIP"
)
}
\value{
character vector available nlTypes
}
\description{
Return the unique nlTypes available in the data
}
\examples{
\dontrun{
 getCtryNlTypes()
}

}
