% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getCtryPolyAdmLevelNames}
\alias{getCtryPolyAdmLevelNames}
\title{Get the list of admin level names in a polygon shapefile}
\usage{
getCtryPolyAdmLevelNames(ctryCode)
}
\arguments{
\item{ctryCode}{the ctryCode of the country of interest}
}
\value{
character vector of admin level names
}
\description{
Get the list of admin level names in a polygon shapefile
}
\examples{
\dontrun{
Rnightlights:::getCtryPolyAdmLevelNames("KEN")
#returns vector [1] "County"       "Constituency" "Ward"
#if KEN shapefile exists otherwise errors
}

}
