% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getCtryRasterOutputFname}
\alias{getCtryRasterOutputFname}
\title{Get the full path to the file where the cropped VIIRS country raster is stored.}
\usage{
getCtryRasterOutputFname(ctryCode, nlYearMonth)
}
\arguments{
\item{ctryCode}{}

\item{nlYearMonth}{the year and month}
}
\value{
Character full path to the cropped VIIRS country raster for a country and a given year and month
}
\description{
Get the full path to the file where the cropped VIIRS country raster is stored. This file is created
    when processing the country before extracting the data. It can be used to re-process a country much faster
}
\examples{
getCtryRasterOutputFname("KEN","201412")

}

