% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appfuns.R
\name{getCtrysDataDirect}
\alias{getCtrysDataDirect}
\title{Read the ctryNlData directly from files}
\usage{
getCtrysDataDirect(
  ctryCodes,
  admLevel,
  gadmPolySrc = "GADM",
  gadmPolyVer = "3.6",
  gadmPolyType = "SHPZIP",
  nlType = "VIIRS.M",
  custPolyPath = NULL
)
}
\arguments{
\item{ctryCodes}{the ISO3 code of the countries}

\item{admLevel}{The country admin level of interest}

\item{gadmPolySrc}{The polygon source to use}

\item{gadmPolyVer}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{nlType}{The nlType of interest}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
character vector available nlTypes
}
\description{
Read the ctryNlData directly from files
}
\examples{
\dontrun{
 getCtryNlTypes()
}

}
