% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getCtryShpAllAdmLvls}
\alias{getCtryShpAllAdmLvls}
\title{Get all the admLevels in a country}
\usage{
getCtryShpAllAdmLvls(ctryCodes)
}
\arguments{
\item{ctryCodes}{\code{character} The ctryCode of the country of interest}
}
\value{
\code{logical} whether inputted admLevels are valid
}
\description{
Get all the admLevels in a country
}
\examples{
\dontrun{
getCtryShpAllAdmLvls("KEN")
#returns "KEN_adm1"
}

}
