% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getCtryShpLowestLyrName}
\alias{getCtryShpLowestLyrName}
\title{Get the name of the lowest ctry admin level}
\usage{
getCtryShpLowestLyrName(ctryCode)
}
\arguments{
\item{ctryCode}{the ctryCode of interest}
}
\value{
character string The name of the lowest admin level
}
\description{
Get the name of the lowest ctry admin level
}
