% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getCtryShpLowestLyrNames}
\alias{getCtryShpLowestLyrNames}
\title{Get the name of the lowest ctry admin level}
\usage{
getCtryShpLowestLyrNames(ctryCodes)
}
\arguments{
\item{ctryCodes}{\code{character} ctryCodes the ctryCodes of interest}
}
\value{
character string The name of the lowest admin level
}
\description{
Get the name of the lowest ctry admin level
}
