% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getCtryStructFname}
\alias{getCtryStructFname}
\title{Construct the name for the country struct file}
\usage{
getCtryStructFname(ctryCode)
}
\arguments{
\item{ctryCode}{The ISO3 ctryCode of the country}
}
\value{
character string The filename
}
\description{
Construct the name for the country struct file
}
\examples{
Rnightlights:::getCtryStructFname("KEN")

}
