% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getCtryStructFnamePath}
\alias{getCtryStructFnamePath}
\title{Construct the full path to the country struct file}
\usage{
getCtryStructFnamePath(ctryCode)
}
\arguments{
\item{ctryCode}{The ISO3 ctryCode of the country}
}
\value{
character string The file path
}
\description{
Construct the full path to the country struct file
}
\examples{
Rnightlights:::getCtryStructFnamePath("KEN")

}
