% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getDataPath}
\alias{getDataPath}
\title{Gets the root path to the file directory"}
\usage{
getDataPath()
}
\arguments{
\item{defaultPath}{The default path, if no user-specified directory
has been given.}

\item{...}{Not used.}
}
\value{
Returns the path as a @character string.
}
\description{
Gets the root path to the file directory"
}
\examples{
  print(getDataPath())

}
\seealso{
To set the directory where package data files are stored,
    see @see "setDataPath".
}

