% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFreeRAM}
\alias{getFreeRAM}
\title{Detect the amount of available RAM on the computer}
\usage{
getFreeRAM()
}
\description{
Detect the amount of available RAM on the computer for dynamic
    configuration of available memory
}
