% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasflares.R
\name{getNlCtryGasFlaresPoly}
\alias{getNlCtryGasFlaresPoly}
\title{Returns a country's gasflares polygon if it exists}
\usage{
getNlCtryGasFlaresPoly(
  ctryCode = ctryCode,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"),
  custPolyPath = NULL
)
}
\arguments{
\item{ctryCode}{\code{character} The ctryCode of the country of interest}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
\code{spPolygons} the country gasflare polygon
}
\description{
Returns a country's gasflares polygon if it exists
}
\examples{
\dontrun{
  getNlCtryGasFlaresPoly(ctryCode="NGA")
  #returns gasflare polygon
}

}
