% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasflares.R
\name{getNlCtryGasFlaresRds}
\alias{getNlCtryGasFlaresRds}
\title{Creates and returns a country's gasflares polygon}
\usage{
getNlCtryGasFlaresRds(ctryCode = ctryCode,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"), custPolyPath = NULL)
}
\arguments{
\item{ctryCode}{\code{character} The ctryCode of the country of interest}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
\code{spPolygons} the country gasflare polygon
}
\description{
Creates and returns a country's gasflares polygon
}
\examples{
\dontrun{
  getNlCtryGasFlaresRds(ctryCode="NGA")
#returns gasflare polygon
}

}
