% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasflares.R
\name{getNlCtryGasFlaresRdsFnamePath}
\alias{getNlCtryGasFlaresRdsFnamePath}
\title{Get the path to the saved country gasflare RDS}
\usage{
getNlCtryGasFlaresRdsFnamePath(ctryCode)
}
\arguments{
\item{ctryCode}{\code{character} The ctryCode of the country of interest}
}
\value{
\code{character} the path to the world gasflare RDS
}
\description{
Get the path to the saved country gasflare RDS
}
\examples{
\dontrun{
  getNlCtryGasFlaresRdsFnamePath(ctryCode="NGA")
  #returns path to the world RDS
}

}
