% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapath.R
\name{getNlDataPath}
\alias{getNlDataPath}
\title{Gets the root path to the file directory"}
\usage{
getNlDataPath()
}
\value{
Returns the folder containing the root of the current data path
    as a @character string.
}
\description{
Gets the root path to the file directory"
}
\examples{
  print(getNlDataPath())

}
\seealso{
To set the directory where package data files are stored,
    see @see "setNlDataPath".
}
