% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapath.R
\name{getNlDataPathFull}
\alias{getNlDataPathFull}
\title{Gets the full path to the data directory"}
\usage{
getNlDataPathFull()
}
\value{
Returns the full folder of the current data path
    as a @character string.
}
\description{
Gets the full path to the data directory"
}
\examples{
print(getNlDataPathFull())
    
}
