% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getNlDir}
\alias{getNlDir}
\title{Get the paths to the various data locations}
\usage{
getNlDir(dirName)
}
\description{
Get the paths to the various locations of nightlights data generated by
    the Rnightlights package. These correspond to the various "dir..." 
    options in the pkgOptions settings
}
