% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapath.R
\name{getNlDirNames}
\alias{getNlDirNames}
\title{Get the names of all possible getNlDir dirNames}
\usage{
getNlDirNames()
}
\description{
Get the names of all possible getNlDir dirNames that can be supplied
    to the getNlDir function as the dirName parameter
}
\examples{
getNlDirNames()

}
