% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasflares.R
\name{getNlGasFlaresRds}
\alias{getNlGasFlaresRds}
\title{Returns the world gasflares polygon if it exists}
\usage{
getNlGasFlaresRds()
}
\value{
\code{spPolygons} the country gasflare polygon
}
\description{
Returns the world gasflares polygon if it exists
}
\examples{
\dontrun{
  getNlGasFlaresRds()
  #returns mosaiced gasflare polygon
}

}
