% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasflares.R
\name{getNlGasFlaresRdsFnamePath}
\alias{getNlGasFlaresRdsFnamePath}
\title{Get the path to the saved world gasflare RDS}
\usage{
getNlGasFlaresRdsFnamePath()
}
\value{
\code{character} the path to the world gasflare RDS
}
\description{
Get the path to the saved world gasflare RDS
}
\examples{
\dontrun{
  getNlGasFlaresRdsFnamePath(ctryCode="NGA")
  #returns path to the world RDS
}

}
