% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTifLclNameOLS}
\alias{getNlTifLclNameOLS}
\title{Constructs the filename used to save/access the decompressed OLS .tif file}
\usage{
getNlTifLclNameOLS(nlYear)
}
\arguments{
\item{nlYear}{the year in which the tile was created}
}
\value{
a character vector filename of the .tif VIIRS tile
}
\description{
Constructs the filename used to save/access the decompressed OLS .tif file
}
\examples{
#using default dirNlTiles
\dontrun{getNlTifLclNameOLS("2004")}
 #returns "ols_2004.tif"


}
