% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getNlTiles}
\alias{getNlTiles}
\title{Create mapping of VIIRS nightlight tiles}
\usage{
getNlTiles()
}
\value{
A data.frame of names of tiles and lon-lat coordinate of top-left corner of each
}
\description{
Creates a data.frame mapping VIIRS nightlight tile names to their vertice coordinates. This is used to
    identify nightlight tiles as well as to build a spatial polygons dataframe used to plot the tiles. OLS
    only has one tile for the whole world and thus does not need mapping.
}
\examples{
getNlTiles()

}
