% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTileTifLclNameOLS}
\alias{getNlTileTifLclNameOLS}
\title{Constructs the filename of the decompressed OLS .tif file}
\usage{
getNlTileTifLclNameOLS(nlPeriod)
}
\arguments{
\item{nlPeriod}{the nlPeriod in which the tile was created}
}
\value{
a character vector filename of the .tif OLS tile
}
\description{
Constructs the filename of the decompressed OLS .tif file
}
\examples{
#using default dirNlTiles
\dontrun{
Rnightlights:::getNlTileTifLclNameOLS("2004")
 #returns "OLS_2004_00N180W.tif"
 }

}
