% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTileTifLclNamePath}
\alias{getNlTileTifLclNamePath}
\title{Constructs the full path used to save/access the downloaded tile .tgz file}
\usage{
getNlTileTifLclNamePath(nlType, nlPeriod, tileNum)
}
\arguments{
\item{nlType}{the nlType of interest}

\item{nlPeriod}{the nlPeriod in which the tile was created}

\item{tileNum}{the index of the tile as given in nlTileIndex}
}
\value{
a character string filename of the compressed .tgz VIIRS tile
}
\description{
Constructs the full path used to save/access the downloaded tile .tgz file
}
\examples{
\dontrun{getNlTileZipLclNamePath("OLS", "2014")}
 #returns "/dataPath/ols_2014_01.tgz"

\dontrun{getNlTileZipLclNamePath("VIIRS", "201412", "1")}
 #returns "/dataPath/viirs_2014_12_75N180W.tgz"
 
}
