% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTileTifLclNamePathOLS}
\alias{getNlTileTifLclNamePathOLS}
\title{Constructs the full path used to save/access the decompressed OLS .tif file}
\usage{
getNlTileTifLclNamePathOLS(nlYear, tileNum)
}
\arguments{
\item{nlYear}{the year in which the tile was created}

\item{tileNum}{ignored}
}
\value{
a character vector filename of the .tif OLS tile
}
\description{
Constructs the full path used to save/access the decompressed OLS .tif file
}
\examples{
#using default dirNlTiles
\dontrun{getNlTileTifLclNamePathOLS("2014", "1")}
 #returns "/dataPath/tiles/ols_2014.tif"

}
