% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTileTifLclNameVIIRS}
\alias{getNlTileTifLclNameVIIRS}
\title{Constructs the filename of the decompressed VIIRS .tif file}
\usage{
getNlTileTifLclNameVIIRS(nlYearMonth, tileNum)
}
\arguments{
\item{nlYearMonth}{the nlYearMonth in which the tile was created}

\item{tileNum}{the index of the tile as given in nlTileIndex}
}
\value{
a character vector filename of the .tif VIIRS tile
}
\description{
Constructs the filename of the decompressed VIIRS .tif file
}
\examples{
#using default dirNlTiles
\dontrun{getNlTileTifLclNameVIIRS("201401", "1")}
 #returns "viirs_2014_01_75N180W.tif"

}
