% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTileZipLclNameOLS}
\alias{getNlTileZipLclNameOLS}
\title{The name with which to save the OLS tile locally}
\usage{
getNlTileZipLclNameOLS(nlYear)
}
\arguments{
\item{nlYear}{The year of the OLS tile}
}
\value{
character string filename
}
\description{
The name with which to save the OLS tile locally
}
\examples{
\dontrun{getNlTileZipLclNameOLS("2012")}

}
