% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTileZipLclNameVIIRS}
\alias{getNlTileZipLclNameVIIRS}
\title{Constructs the filename used to save/access the downloaded VIIRS tile .tgz file}
\usage{
getNlTileZipLclNameVIIRS(nlPeriod, tileNum, nlType)
}
\arguments{
\item{nlPeriod}{The nlPeriod in which the tile was created}

\item{tileNum}{The index of the tile as given in nlTileIndex}

\item{nlType}{character the nlType}
}
\value{
A character string filename of the compressed .tgz VIIRS tile
}
\description{
Constructs the filename used to save/access the downloaded VIIRS tile .tgz file
}
\examples{
\dontrun{
Rnightlights:::getNlTileZipLclNameVIIRS("201401", 1)
 #returns "./tiles/VIIRS_2014_01_75N180W.tgz"
 }

}
