% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlurl.R
\name{getNlUrl}
\alias{getNlUrl}
\title{Function to return the url of the tile to download}
\usage{
getNlUrl(nlPeriod)
}
\arguments{
\item{nlPeriod}{The nlPeriod of the tile for which to return the tile download URL}
}
\value{
character string Url of the tile file
}
\description{
Function to return the url of the tile to download given the year
}
\examples{
\dontrun{
tileUrl <- Rnightlights:::getNlUrlOLS("1999")
}

}
