% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlurl.R
\name{getNlUrlVIIRS}
\alias{getNlUrlVIIRS}
\title{Function to return the url of the VIIRS tile to download}
\usage{
getNlUrlVIIRS(nlYearMonth, tileNum)
}
\arguments{
\item{nlYearMonth}{character string yearmonth in "YYYYMM' format e.g. "201401"}

\item{tileNum}{The integer index of the tile to download as given by getNlTiles("VIIRS")}
}
\value{
Character string Url of the VIIRS tile file
}
\description{
Function to return the url of the VIIRS tile to download given the year, month, and nlTile index
}
\examples{
\dontrun{tileUrl <- getNlUrlVIIRS("201401", "1")}

}
