% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{getNlYearMonthTilesVIIRS}
\alias{getNlYearMonthTilesVIIRS}
\title{Download the listed VIIRS tiles for a particular nlYearMonth}
\usage{
getNlYearMonthTilesVIIRS(nlYearMonth, tileList)
}
\arguments{
\item{nlYearMonth}{character the nlYearMonth to process in the format "YYYYMM"}

\item{tileList}{integer vector or character vector of digits containing valid tile numbers}
}
\value{
TRUE/FALSE if the download was successful
}
\description{
Download the listed VIIRS tiles for a particular nlYearMonth
}
\examples{
#download tiles for "KEN" which are tiles 2 and 5 for the specified time periods
getNlYearMonthsTilesVIIRS("201201", c(2, 5))

#same as above but getting the tileList automatically
getNlYearMonthsTilesVIIRS("201201", tileName2Idx(getCtryCodeTileList("KEN")))

returns TRUE if the download was successful

}

