% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getNtLtsOLS}
\alias{getNtLtsOLS}
\title{Download OLS nightlight tile in TIF format from NOAA compressed as .tgz}
\usage{
getNtLtsOLS(nlYear, tileNum)
}
\arguments{
\item{nlYear}{the year in "YYYY" format e.g. "2012"}
}
\value{
TRUE/FALSE Whether the download was successful
}
\description{
Download OLS nightlight tile in TIF format from NOAA compressed as .tgz
}
\examples{
result <- getNtLtsVIIRS("2012", "05", "1")

if (result)
  print("download successful")

}

