% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getNtLtsTifLclNameOLS}
\alias{getNtLtsTifLclNameOLS}
\title{Constructs the filename used to save/access the decompressed OLS .tif file}
\usage{
getNtLtsTifLclNameOLS(nlYear, tileNum)
}
\arguments{
\item{nlYear}{the year in which the tile was created}

\item{tileNum}{the index of the tile as given in nlTileIndex}

\item{nlMonth}{the month in which the tile was created}

\item{dir}{the directory storing the tiles. Defaults to the global variable dirRasterVIIRS}
}
\value{
a character vector filename of the .tif VIIRS tile
}
\description{
Constructs the filename used to save/access the decompressed OLS .tif file
}
\examples{
#using default dirRasterOLS
getNtLtsTifLclNameOLS("2014", "01", "1")
 returns "./tiles/viirs_2014_01_75N180W.tif"


}

