% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getNtLtsUrlOls}
\alias{getNtLtsUrlOls}
\title{Function to return the url of the OLS tile to download}
\usage{
getNtLtsUrlOls(inYear, inTile)
}
\arguments{
\item{inYear}{}
}
\value{
Character string Url of the VIIRS tile file
}
\description{
Function to return the url of the OLS tile to download given the year, month, and nlTile index
}
\examples{
tileUrl <- getNtLtsUrlOls("1999")

}

