% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{getNtLtsUrlVIIRS}
\alias{getNtLtsUrlVIIRS}
\title{Function to return the url of the VIIRS tile to download}
\usage{
getNtLtsUrlVIIRS(inYear, inMonth, tileNum)
}
\arguments{
\item{inYear}{}

\item{inMonth}{character month in MM format e.g. Jan="01", Feb="02"}

\item{tileNum}{The integer index of the tile to download as given by getNlTiles()}
}
\value{
Character string Url of the VIIRS tile file
}
\description{
Function to return the url of the VIIRS tile to download given the year, month, and nlTile index
}
\examples{
tileUrl <- getNtLtsUrlVIIRS("2012", "04", "1")

}

