% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{getNtLtsVIIRS}
\alias{getNtLtsVIIRS}
\title{Download VIIRS nightlight tile}
\usage{
getNtLtsVIIRS(nlYear, nlMonth, tileNum,
  downloadMethod = pkg_options("downloadMethod"))
}
\arguments{
\item{nlYear}{the year in "YYYY" format e.g. "2012"}

\item{tileNum}{the index of the tile as given by getNlTiles()}

\item{monthNum}{the month in "MM" format e.g. Jan="01", Feb="02"}
}
\value{
TRUE/FALSE Whether the download was successful
}
\description{
Download VIIRS nightlight tile
}
\examples{
if(getNtLtsVIIRS("2012", "05", "1"))
  print("download successful")

}

