% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{getNtLtsZipLclNameVIIRS}
\alias{getNtLtsZipLclNameVIIRS}
\title{Constructs the filename used to save/access the downloaded VIIRS tile .tgz file}
\usage{
getNtLtsZipLclNameVIIRS(nlYear, nlMonth, tileNum,
  dir = pkg_options("dirRasterVIIRS"))
}
\arguments{
\item{nlYear}{the year in which the tile was created}

\item{nlMonth}{the month in which the tile was created}

\item{tileNum}{the index of the tile as given in nlTileIndex}

\item{dir}{the directory storing the tiles. Defaults to the global variable dirRasterVIIRS}
}
\value{
a character string filename of the compressed .tgz VIIRS tile
}
\description{
Constructs the filename used to save/access the downloaded VIIRS tile .tgz file
}
\examples{
getNtLtsZipLclNameVIIRS("2014", "01", "1")
 returns "./tiles/viirs_2014_01_75N180W.tgz"

}

