% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getSavedNlStat}
\alias{getSavedNlStat}
\title{Retrieve a saved nlStat function}
\usage{
getSavedNlStat(nlStatSignature)
}
\arguments{
\item{nlStatSignature}{character The signature of the nlStat to retrieve which
is used as the key of the saved nlStats}
}
\value{
list A list representing the saved nlStat
}
\description{
Retrieve a saved nlStat function
}
\examples{

Rnightlights:::getSavedNlStat("sum()")

}
