% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getSavedNlStatFname}
\alias{getSavedNlStatFname}
\title{The name of the file in which the saved nlStats are stored}
\usage{
getSavedNlStatFname()
}
\value{
character The name of the saved nlStats file
}
\description{
The name of the file in which the saved nlStats are stored
}
