% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getSavedNlStatFnamePath}
\alias{getSavedNlStatFnamePath}
\title{The path tos the file in which the saved nlStats are stored}
\usage{
getSavedNlStatFnamePath()
}
\value{
character The path to the saved nlStats file
}
\description{
The path to the file in which the saved nlStats are stored
}
