% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{getTilesCtryIntersectVIIRS}
\alias{getTilesCtryIntersectVIIRS}
\title{Get a list of tiles that a country polygon intersects with}
\usage{
getTilesCtryIntersectVIIRS(ctryCode)
}
\arguments{
\item{ctryCode}{The country's ISO3 code}
}
\value{
None
}
\description{
Create a dataframe mapping each country in the rworldmap to the VIIRS
    tiles which they intersect with and thus need to be retrieved to
    process their nightlight imagery. Since some functions use this
    dataframe for long-term processing, omitCountries can eliminate 
    countries that should be excluded from the list hence from processing.
    Countries can be added in the omitCountries function.
    Default is "none".
}
\examples{

Rnightlights:::getTilesCtryIntersectVIIRS("KEN")

}
