% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getWorldMap}
\alias{getWorldMap}
\title{Make the rworldmap available for other functions}
\usage{
getWorldMap()
}
\description{
Make the rworldmap available for other functions. Since running cleangeo
    takes some time we want to run it once the first time it is called
    and make it available to other functions without overhead. The map
    is stored to the .RnightlightsEnv hidden package environment
}
