% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{Run a web application to explore the processed nightlight data cached
    locally}
\usage{
gui()
}
\value{
None
}
\description{
Run a web application to perform some exploratory data analysis. The
    application written in shiny either loads demo data or the data 
    processed and cached locally.
}
\examples{
\dontrun{
gui()
}

}
