% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasflares.R
\name{hasNlCtryGasFlares}
\alias{hasNlCtryGasFlares}
\title{Checks if a country has gasflares}
\usage{
hasNlCtryGasFlares(ctryCode, gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"), custPolyPath = NULL)
}
\arguments{
\item{ctryCode}{\code{character} The ctryCode of the country of interest}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
\code{logical} whether the country has gasflares
}
\description{
Checks if a country has gasflares
}
\examples{
\dontrun{
  hasNlCtryGasFlares(ctryCode="NGA")
#returns TRUE
}

}
