% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{hCluster}
\alias{hCluster}
\title{Plots the ctryNlData}
\usage{
hCluster(
  ctryData,
  ctryCodes,
  admLevel,
  nlType,
  nlStat,
  nlPeriodStart,
  nlPeriodEnd,
  graphType,
  normArea
)
}
\arguments{
\item{ctryData}{data.frame If provided the plot will use data from it rather 
than retrieving it. If not available the follow parameters are required
to retrieve the data using getCtryNlData.

It is assumed that the ctryData data.frame is in the same format as the
output from getCtryNlData i.e. admin cols first, an area column,
followed by the stat columns. In future it may be possible to supply
a custom data.frame and direct the function on which columns to find
the particular data.}

\item{ctryCodes}{\code{character} The ctryCode of interest}

\item{admLevel}{\code{character} The country admin level in the given
ctryCode at which to calculate stats}

\item{nlType}{\code{character} The nlType of interest}

\item{nlStat}{the statistics to calculate. If not provided will calculate
the stats specified in \code{pkgOptions("nlStats")}}

\item{configName}{character the config shortname of raster being processed}

\item{extension}{character the extension of raster being processed}

\item{removeGasFlaresMethod}{character The method to use to perform gas flare removal
or NULL to disable}

\item{nlPeriod}{\code{character} The nlPeriod of interest}

\item{downloadMethod}{The method used to download polygons and rasters}

\item{cropMaskMethod}{\code{character} Whether to use rasterize or
gdal-based functions to crop and mask the country rasters}

\item{extractMethod}{("rast" or "gdal") Whether to use rasterize or
gdal-based functions to crop and mask the country rasters}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
None
}
\description{
Plots the ctryNlData in various plot styles including line, box
    and density plot using ggplot2. If ctryData is not provided the subsequent
    parameters will need to be supplied so thta getCtryNlData can be used to 
    retrieve the data prior to plotting.
}
\examples{

#calculate only the sum of monthly VIIRS radiances for Dec 2014 using gdal
#for both cropMask and extraction for KEN
\dontrun{
Rnightlights:::processNLCountry("KEN", "KEN_adm2", "VIIRS.M", "201412", "gdal", "gdal", "sum")
}

}
