% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{initNtLts}
\alias{initNtLts}
\title{Initialize some important variables and create directory structure}
\usage{
initNtLts(omitCountries = "none")
}
\description{
Initialize some important variables and create directory structure
}
\examples{

 initNtLts()

}
