% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{listSavedNlStats}
\alias{listSavedNlStats}
\title{List saved nlStats}
\usage{
listSavedNlStats(nlStatNames = NULL, detail = FALSE)
}
\arguments{
\item{nlStatNames}{character The signatures of the nlStats to retrieve}

\item{detail}{logical Whether to print out the whole saved nlStat including
the signature, body, arguments and hash}
}
\value{
list A list of lists representing the saved nlStats
}
\description{
List saved nlStats
}
\examples{
listSavedNlStats("sum()")

#returns
#"sum()"


listSavedNlStats("sum()", detail = TRUE)

#returns
#$`sum()`
#$`sum()`$nlStatBody
#function (..., na.rm = FALSE)  .Primitive("sum")
#
#$`sum()`$nlStatArgs
#[1] ""
#
#$`sum()`$nlStatHash
#[1] "f0fbe35d81578311ba8f362137832e779b7b4f39"

}
