% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{loadSavedNlStat}
\alias{loadSavedNlStat}
\title{Load the saved nlStat into an accessible environment}
\usage{
loadSavedNlStat(
  statSig,
  newStatName = NULL,
  envir = .GlobalEnv,
  overwrite = FALSE
)
}
\arguments{
\item{statSig}{character The nlStat signature}

\item{newStatName}{character The name to give the nlStat in the environment
in case one does not want to use the name it is saved with}

\item{envir}{character The environment in which to append the nlStat}

\item{overwrite}{logical Whether to overwrite a function with the same name
if it is present}
}
\value{
logical Whether the process was successful
}
\description{
Load the saved nlStat into an accessible environment e.g. the global
    environment. This is especially required if the function is saved
    but not present in the Global environment or an environment where
    the package functions can access it
}
\examples{
\dontrun{
  loadSavedNlStat(statSig = "sum()", newStatName = "mySum", overwrite = TRUE)
}

#returns "sum"

}
