% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{mapAllCtryPolyToTiles}
\alias{mapAllCtryPolyToTiles}
\title{Create a mapping of all countries and the tiles they intersect}
\usage{
mapAllCtryPolyToTiles(omitCountries = pkg_options("omitCountries"))
}
\arguments{
\item{omitCountries}{A character vector or list of countries to leave out when processing. Default is #'     \code{"none"}}
}
\value{
None
}
\description{
This is simply another name for mapCtryPolyToTiles with ctryCodes="all"
}
\examples{
mapAllCtryPolyToTiles() no countries omitted

mapAllCtryPolyToTiles(omitCountries="none") no countries omitted

mapAllCtryPolyToTiles(omitCountries=c("error", "long")) will not omit countries that take long to process

}

