% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{mapAllCtryPolyToTilesVIIRS}
\alias{mapAllCtryPolyToTilesVIIRS}
\title{Create a mapping of all countries and the tiles they intersect}
\usage{
mapAllCtryPolyToTilesVIIRS(omitCountries = pkgOptions("omitCountries"))
}
\arguments{
\item{omitCountries}{A character vector or list of countries to leave
out when processing. Default is \code{"none"}}
}
\value{
None
}
\description{
This is simply another name for mapCtryPolyToTilesVIIRS with ctryCodes="all"
}
\examples{
#no countries omitted
\dontrun{
tileMap <- Rnightlights:::mapAllCtryPolyToTilesVIIRS()
}

#no countries omitted
\dontrun{
tileMap <- Rnightlights:::mapAllCtryPolyToTilesVIIRS(omitCountries="none")
}

#include countries that take long to process
\dontrun{
tileMap <- Rnightlights:::mapAllCtryPolyToTilesVIIRS(omitCountries=c("error", "long"))
}

}
