% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masq.R
\name{masqOLS}
\alias{masqOLS}
\title{Extract raster pixel values within the boundaries of a polygon}
\usage{
masqOLS(ctryPoly, ctryRast, idx, retVal = NULL, configName, extension)
}
\arguments{
\item{ctryPoly}{the country Polygon layer as SpatialPolygon}

\item{ctryRast}{the clipped country raster}

\item{idx}{the index of the polygon in the country polygon layer (shp)}

\item{retVal}{Whether to return the raster data as a vector, or
data.frame with spatial context NULL returns a vector of all
values, colrowval returns a data.frame with row, col and raster
value while lonlatval returns a data.frame with lon,lat and val.}

\item{configName}{character the config shortname of raster being processed}

\item{extension}{character the extension of raster being processed}
}
\value{
numeric vector of radiances
}
\description{
Extract raster pixel values within the boundaries of a polygon
}
\examples{
\dontrun{
ctryPoly <- rgdal::readOGR(getPolyFnamePath("KEN"), getCtryShpLyrNames("KEN", 1))
ctryRaster <- raster::raster(getCtryRasterOutputFnamePath("KEN", "OLS", "1999"))
temp <- NULL
KenAdm1Sum <- NULL
for (i in 1:length(ctryPoly@polygons))
{
 temp$name <- as.character(ctryPoly@data$NAME_1[i])
 temp$sum <- sum(masqOLS(ctryPoly, ctryRaster, i), na.rm=T)

 KenAdm1Sum <- rbind(KenAdm1Sum)
}
}

}
