% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{maxPtsNeighborDistToHalfThresh}
\alias{maxPtsNeighborDistToHalfThresh}
\title{Function to test the use of rowcols and lonlats in getCtryNlData}
\usage{
maxPtsNeighborDistToHalfThresh(
  col,
  row,
  lon,
  lat,
  val,
  numPts = 3,
  enablePlot = FALSE
)
}
\arguments{
\item{col}{character parameter to tell the package to pass in the colrow
values i.e. column and row indices as per the country polygon}

\item{row}{character parameter to tell the package to pass in the colrow
values i.e. column and row indices of non-NA pixels as per the country
polygon}

\item{lon}{character parameter to tell the package to pass in the lonlat
values i.e. column and row indices of non-NA pixels as per the country
polygon}

\item{lat}{character parameter to tell the package to pass in the lonlat
values i.e. column and row indices of non-NA pixels as per the country
polygon}

\item{val}{character variable which will hold the raster values found in
in the raster}

\item{numPts}{integer The number of top pixels to detect}

\item{enablePlot}{logical Plot each stage of the iteration showing the
pixels surrounding the top pixel as we move farther out till the
target pixel is found. Output will be in PDF format.}
}
\description{
Function to test the use of rowcols and lonlats in getCtryNlData as
    input to the nlStats parameter.
}
\details{
The function gets the top numPts brightest pixels in a sub-polygon
    or zone and looks at increasing neighbouring distances looking
    for the distance to the first pixel that is 1/2 the value of the
    central pixel.
}
\examples{

\dontrun{
pkgReset();
  system.time(rastGFT92 <-  getCtryNlData(ctryCode = "NGA",
     admLevel = "lowest", nlTypes = "OLS.Y", nlPeriods = "1992",
     removeGasFlaresMethod = "OGP", ignoreMissing=FALSE, 
     nlStats = list("maxPtsNeighborDistToHalfThresh", 
     list(numPts=1, enablePlot=TRUE))))

  pkgOptions(extractMethod="gdal",cropMaskMethod="gdal");
  system.time(gdalGFT92 <-  getCtryNlData(ctryCode = "NGA",
      admLevel = "lowest", nlTypes = "OLS.Y", nlPeriods = "1992",
      removeGasFlaresMethod = "OGP", ignoreMissing=FALSE,
      nlStats = list("maxPtsNeighborDistToHalfThresh", 
      list(numPts=1, enablePlot=TRUE))))
}

}
