% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{myZonal}
\alias{myZonal}
\title{Calculate zonal statistics. Used internally}
\usage{
myZonal(x, z, stats, digits = 0, na.rm = TRUE, ...)
}
\arguments{
\item{x}{the country raster}

\item{z}{the zonal country polygon layer}

\item{stats}{a character list of statistics to calculate}

\item{digits}{round off to how many decimals}

\item{na.rm}{how to handle NAs}

\item{...}{}
}
\value{
numeric value result of the given stat function
}
\description{
Calculate zonal statistics. Used internally by zonalpipe. Modified from 
    \url{http://www.guru-gis.net/efficient-zonal-statistics-using-r-and-gdal/}
}
\examples{
myZonal

}

