% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlInit}
\alias{nlInit}
\title{Initialize some important variables and create directory structure}
\usage{
nlInit(omitCountries = "none")
}
\arguments{
\item{omitCountries}{character string/vector CtryCodes to exclude from processing}
}
\description{
Initialize some important variables and create directory structure
}
\examples{

 \dontrun{
 Rnightlights:::nlInit()
 }

}
