% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{nlPeriodToDate}
\alias{nlPeriodToDate}
\title{Convert nlPeriod to date}
\usage{
nlPeriodToDate(nlPeriod, nlType)
}
\arguments{
\item{nlPeriod}{the nlPeriod to convert}

\item{nlType}{the nlType to use}
}
\value{
a date vector
}
\description{
Convert nlPeriod to date
}
\examples{
nlPeriodToDate(nlPeriod = "201204", nlType = "VIIRS.M")
#returns "2012-04-01"

}
