% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{nlRange}
\alias{nlRange}
\title{Create a range of nlPeriods}
\usage{
nlRange(startNlPeriod, endNlPeriod, nlType)
}
\arguments{
\item{startNlPeriod}{the nlPeriod start}

\item{endNlPeriod}{the nlPeriod end}

\item{nlType}{the nlType}
}
\value{
character vector of nlPeriods
}
\description{
Create a range of nlPeriods. Autodetects the type of nlPeriod and creates a 
    character vector of nlPeriods filling in the intermediate nlPeriods.
    NOTE: Both start and end range must be valid and of the same type i.e.
    "OLS" years or "VIIRS" yearMonths
}
\examples{
#get OLS years between 2004 and 2010
nlRange("2004", "2010", "OLS.Y")

#get VIIRS yearMonths between Jan 2014 and Dec 2014
nlRange("201401", "201412", "VIIRS.M")

}
