% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{nlSignatureAddArg}
\alias{nlSignatureAddArg}
\title{Add an argument to an nlSignature}
\usage{
nlSignatureAddArg(nlStatSigs, addArg)
}
\arguments{
\item{nlStatSigs}{character A vector of nlStat signatures to which to add
the argument}

\item{addArg}{character The argument to add to the signatures}
}
\value{
named character A vector of modified nlStat signatures
}
\description{
Add an argument to an nlSignature
}
\examples{

nlStat <- list("sum")

nlSig <- Rnightlights:::nlStatSignature(nlStat)

Rnightlights:::nlSignatureAddArg(nlStatSigs = nlSig, addArg = "na.rm = TRUE")

#returns
#            sum() 
# "sum(na.rm=TRUE)" 

}
