% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{nlSignatureStat}
\alias{nlSignatureStat}
\title{Convert a signature into an nlStat}
\usage{
nlSignatureStat(nlStatSignature)
}
\arguments{
\item{nlStatSignature}{character The signature of an nlStat}
}
\value{
named character An nlStat
}
\description{
Convert a signature into an nlStat
}
\examples{

nlSignature <- "sum(na.rm=T)"

Rnightlights:::nlSignatureStat(nlStatSignature = nlSignature)

#returns an nlStat list
#[[1]]
#[[1]][[1]]
#[1] "sum"
# 
#[[1]][[2]]
#[1] "na.rm=TRUE"
#
#list structure
#List of 1
#$ :List of 2
#..$ : chr "sum"
#..$ : chr "na.rm=TRUE"

}
