% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{nlSignatureStatName}
\alias{nlSignatureStatName}
\title{get the name of the nlStat from the signature}
\usage{
nlSignatureStatName(statSig)
}
\arguments{
\item{statSig}{character The nlStat signature}
}
\value{
character the name of the nlStat
}
\description{
get the name of the nlStat from the signature i.e. the part before the
    first bracket
}
\examples{
nlSignatureStatName("sum()")

#returns "sum"

}
