% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{nlStatArgs}
\alias{nlStatArgs}
\title{Get the arguments passed to an nlStat}
\usage{
nlStatArgs(nlStat)
}
\arguments{
\item{nlStat}{the name of the nlStat to process}
}
\value{
named character A vector of arguments passed to the nlStat
}
\description{
Get the arguments passed to an nlStat when specifying it
}
\examples{

nlStat <- list("sum", "na.rm=TRUE")

Rnightlights:::nlStatArgs(nlStat = nlStat)

#returns "na.rm=TRUE"

}
