% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{nlStatArgsStandardize}
\alias{nlStatArgsStandardize}
\title{Standardize the arguments of an nlStat}
\usage{
nlStatArgsStandardize(nlStatArg)
}
\arguments{
\item{nlStatArg}{character The argument of an nlStat}
}
\value{
named character The signature of the nlStat
}
\description{
Standardize the arguments of an nlStat to ensure signatures can be
    compared. Currently this entails converting
    logical `T` to TRUE and `F` to FALSE
}
\examples{

Rnightlights:::nlStatArgsStandardize("na.rm=T")

#returns "na.rm=TRUE"

}
