% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{nlStatSignature}
\alias{nlStatSignature}
\title{Get the signature of an nlStat}
\usage{
nlStatSignature(nlStat)
}
\arguments{
\item{nlStat}{character The name of an nlStat}
}
\value{
named character The signature of the nlStat
}
\description{
Get the signature of an nlStat i.e. the nlStat function with all arguments
    passed in by the user. Used to uniquely identify an nlStat function call
    i.e. the same function called with different arguments results in a
    different signature
}
\examples{

nlStat <- list("sum")

nlSig <- Rnightlights:::nlStatSignature(nlStat)

#returns "sum()"

nlStat <- list("sum", "na.rm=TRUE")

nlSig <- Rnightlights:::nlStatSignature(nlStat)

#returns "sum(na.rm=TRUE)"

}
