% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{orderCustPolyLayers}
\alias{orderCustPolyLayers}
\title{Order polygon shapefile layers in custom polygons}
\usage{
orderCustPolyLayers(ctryCode, custPolyPath = NULL)
}
\arguments{
\item{ctryCode}{The ISO3 ctryCode of the country polygon to download}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
None
}
\description{
Order polygon shapefile layers in custom polygons
}
\details{
Add an index column to all layers of a polygon
}
\examples{
\dontrun{
Rnightlights:::addCtryPolyIdx(ctryCode="KEN")
}

}
