% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pkgReset}
\alias{pkgReset}
\title{Reset global options for the Rnightlights package}
\usage{
pkgReset()
}
\description{
Reset global options for the Rnightlights package
}
\examples{
#get cropMaskMethod
pkgOptions("cropMaskMethod")

#set cropMaskMethod to "gdal"
pkgOptions(cropMaskMethod="gdal")

#check cropMaskMethod has changed
pkgOptions("cropMaskMethod")

#reset pkgOptions
pkgReset()

#check cropMaskMethod has been reset
pkgOptions("cropMaskMethod")

}
