% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{prettyNlSignature}
\alias{prettyNlSignature}
\title{Remove triple-dots (...) from names of arguments in a signature}
\usage{
prettyNlSignature(nlStatSig)
}
\arguments{
\item{nlStatSig}{character The signature of an nlStat}
}
\value{
named character An nlStat signature without the triple-dots
}
\description{
Remove triple-dots (...) from names of arguments in a signature
}
\examples{

nlStat <- list("sum", "x=10", "na.rm=T")

nlStatSig <- Rnightlights:::nlStatSignature(nlStat) #"sum(...=x=10,na.rm=TRUE)"

Rnightlights:::prettyNlSignature(nlStatSig = nlStatSig)
#returns "sum(x=10,na.rm=TRUE)"

}
