% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{readSavedNlStats}
\alias{readSavedNlStats}
\title{Read saved nlStats from disk to memory}
\usage{
readSavedNlStats()
}
\value{
none
}
\description{
Read saved nlStats from disk to memory and save them into the Rnightlights
    environment
}
\examples{
\dontrun{
  readSavedNlStats()
}

}
