% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadnltiles.R
\name{reqAuthTokenEOG}
\alias{reqAuthTokenEOG}
\title{Retrieve an access token required for download from the EOG site}
\usage{
reqAuthTokenEOG()
}
\description{
Retrieve a temporary access token required for the actual download from the
   new EOG site at University of Colorado, Dept of Mines
}
