% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadnltiles.R
\name{saveCredentialsEOG}
\alias{saveCredentialsEOG}
\title{Save credentials required for download from the EOG site}
\usage{
saveCredentialsEOG(
  credFile = file.path(getNlDataPathFull(), pkgOptions("EOG_CredFile"))
)
}
\arguments{
\item{credFile}{(character) Path to a file containing the credentials
obtained from registration at 
\url{https://eogauth.mines.edu/auth/realms/master/account/}}
}
\description{
Save credentials required for download from the new EOG site at University of
    Colorado, Dept of Mines
}
