% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{saveCtryNlData}
\alias{saveCtryNlData}
\title{Save a data frame of a country's data to the appropriate location}
\usage{
saveCtryNlData(ctryNlDataDF, ctryCode)
}
\arguments{
\item{ctryNlDataDF}{dataframe with the country data to save}
}
\value{
None
}
\description{
Saves the data frame created from processNlCountry* to the appropriate location. 
    Note: This function does not perform any validation error checking and will overwrite 
    existing data. Use with caution.
}
\examples{

saveCtryNlData(ctryNlDataDF, ctryCode)

}

