% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{savedNlStatsIsLoaded}
\alias{savedNlStatsIsLoaded}
\title{Check if saved nlStats have been loaded into memory from disk}
\usage{
savedNlStatsIsLoaded()
}
\value{
logical If the saved nlStats have been loaded
}
\description{
Check if saved nlStats have been loaded into memory from disk
}
\examples{

Rnightlights:::savedNlStatsIsLoaded()

#returns TRUE/FALSE

}
