% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{saveNlStat}
\alias{saveNlStat}
\title{Remove triple-dots (...) from names of arguments in a signature}
\usage{
saveNlStat(nlStat)
}
\arguments{
\item{nlStat}{character An nlStat}
}
\value{
logical Whether the process was successful
}
\description{
Remove triple-dots (...) from names of arguments in a signature
}
\examples{

nlStat <- list("sum", "na.rm=TRUE")

saveNlStat(nlStat)

}
