% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{searchAdmLevel}
\alias{searchAdmLevel}
\title{Search for the admLevel by official name}
\usage{
searchAdmLevel(ctryCode, admLevelName, dnldPoly)
}
\arguments{
\item{ctryCode}{\code{character} The ctryCode of the country of interest}

\item{admLevelName}{\code{character} The name to search for}

\item{dnldPoly}{\code{logical} If the country polygon doesn't exist 
should we download it?}
}
\value{
character vector of admin level names
}
\description{
Search for the admLevel by official name. Expects the shapefile
    to already exist.
}
\examples{
\dontrun{
searchAdmLevel("KEN", "county")
#returns "KEN_adm1"
}

}
