% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{searchSavedNlStatName}
\alias{searchSavedNlStatName}
\title{Check if an nlStat has been saved using the nlStatName}
\usage{
searchSavedNlStatName(nlStatName)
}
\arguments{
\item{nlStatName}{character The name of the nlStat to search for}
}
\value{
character The signatures of nlStat that match or NULL if not found
}
\description{
Check if an nlStat has been saved using the nlStatName
}
\examples{
nlSignatureStatName("sum()")

#returns "sum"

}
