% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{setDataPath}
\alias{setDataPath}
\title{Sets the root path to the package data directory}
\usage{
setDataPath(dataPath, newDataPath)
}
\arguments{
\item{path}{The path}

\item{...}{Not used}
}
\value{
Returns (invisibly) the old root path.
}
\description{
By default, this function will set the root path to \code{~/.Rnightlights/}.
}
\examples{
  
}

