% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapath.R
\name{setDataVersion}
\alias{setDataVersion}
\title{Add README file to the root data path}
\usage{
setDataVersion(path = getNlDataPath(), version)
}
\arguments{
\item{to}{The folder to add the README file to}
}
\value{
None
}
\description{
Add README file to the root data path
}
\examples{
  \dontrun{
  Rnightlights:::addREADME()
  }
    
}
