% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{setupDataPath}
\alias{setupDataPath}
\title{Interactively offers the user to set up the default root path}
\usage{
setupDataPath(newDataPath = "~", ...)
}
\arguments{
\item{...}{Not used.}

\item{defaultPath}{Default root path to set}
}
\value{
character string Returns (invisibly) the root path,
    or @NULL if running a non-interactive session.
}
\description{
Interactively offers the user to set up the default root path
}
\seealso{
Internally, @see "setDataPath" is used to set the root path.
    The \code{"base::interactive"} function is used to test whether 
    \code{R} is running interactively or not.
}

