% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{tileIdx2Name}
\alias{tileIdx2Name}
\title{Get the name of a tile given its index}
\usage{
tileIdx2Name(tileNum)
}
\arguments{
\item{tile}{index as given by getNlTiles()}
}
\value{
Character name of the tile
}
\description{
Get the name of a VIIRS tile as given by getNlTiles() given its index
}
\examples{
tileIdx <- tileName2Idx("00N060W")

}

