% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{tileName2Idx}
\alias{tileName2Idx}
\title{Get the index of a tile given its name}
\usage{
tileName2Idx(tileName)
}
\arguments{
\item{tile}{name as given by getNlTiles()}
}
\value{
Integer index of the tile
}
\description{
Get the index of a VIIRS tile as given by getNlTiles() given its name
}
\examples{
tileIdx <- tileName2Idx("00N060W")

}
