% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{tilesPolygonIntersect}
\alias{tilesPolygonIntersect}
\title{Get the list of VIIRS tiles that a polygon intersects with}
\usage{
tilesPolygonIntersect(shpPolygon)
}
\arguments{
\item{a}{SpatialPolygon or SpatialPolygons}
}
\value{
Character vector of the intersecting tiles as given by getNlTiles()
}
\description{
Get the list a VIIRS tiles that a polygon intersects with
}
\examples{
ctryShapefile <- dnldCtryPoly("KEN")
ctryPoly <- rgdal::readOGR(getPolyFnamePath("KEN"), getCtryShpLyrName("KEN",0))
tileList <- tilesPolygonIntersect(ctryPoly)

}

