% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{tilesPolygonIntersectVIIRS}
\alias{tilesPolygonIntersectVIIRS}
\title{Get the list of VIIRS tiles that a polygon intersects with}
\usage{
tilesPolygonIntersectVIIRS(shpPolygon)
}
\arguments{
\item{shpPolygon}{a SpatialPolygon or SpatialPolygons}
}
\value{
Character vector of the intersecting tiles as given by getNlTiles("VIIRS")
}
\description{
Get the list a VIIRS tiles that a polygon intersects with
}
\examples{
\dontrun{
ctryShapefile <- dnldCtryPoly("KEN")
ctryPoly <- rgdal::readOGR(getPolyFnamePath("KEN"), getCtryShpLyrName("KEN",0))
tileList <- tilesPolygonIntersectVIIRS(ctryPoly)
}

}
