% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgradeRnightlights}
\alias{upgradeRnightlights}
\title{Perform upgrade functions to new package versions as required}
\usage{
upgradeRnightlights()
}
\value{
TRUE/FALSE
}
\description{
Perform upgrade functions to new package versions as required
    General Upgrade functions
    pre-0.2.0 to 0.2.0:
    \itemize{
        \item{Rename tiles}
        \item{Rename output rasters}
        \item{Rename data files}
        \item{Rename data column names}
        \item{Remove zonal rasters which will be regenerated when required}
    }
}
\examples{
\dontrun{
  Rnightlights:::upgradeRnightlights()
  #returns TRUE/FALSE
}
}
