% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{validCtryCode}
\alias{validCtryCode}
\title{Check if a month number is valid for a given nightlight type}
\usage{
validCtryCode(ctryCode)
}
\arguments{
\item{ctryCode}{the ISO3 country code to validate}
}
\value{
TRUE/FALSE
}
\description{
Check if a month number is valid for a given nightlight type. Note month num is only valid for
"VIIRS" nightlight type
}
\examples{
validNlCtryCode("KEN")
 #returns TRUE

validNlCtryCode("UAE")
 #returns FALSE. "United Arab Emirates" ISO3 code = "ARE"

}
