% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{validCtryNlDataDF}
\alias{validCtryNlDataDF}
\title{Check if a country dataframe is valid}
\usage{
validCtryNlDataDF(ctryNlDataDF)
}
\arguments{
\item{ctryNlDataDF}{the country dataframe}
}
\value{
TRUE/FALSE
}
\description{
Check if a country dataframe is valid
}
\examples{
validCtryNlDataDF(nlCtryDataDF)
 returns TRUE

validNlMonthName("","OLS")
 returns FALSE

}

