% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{validGadmPolyTypes}
\alias{validGadmPolyTypes}
\title{Check whether GADM polyTypes are valid}
\usage{
validGadmPolyTypes(gadmPolyTypes)
}
\arguments{
\item{gadmPolyTypes}{The GADM polyTypes to validate}
}
\value{
logical vector
}
\description{
Check whether GADM polyTypes are valid
}
\examples{
\dontrun{
Rnightlights:::validGadmPolyTypes("shpZip")
}

}
