% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{validGadmVersions}
\alias{validGadmVersions}
\title{Check whether GADM versions are valid}
\usage{
validGadmVersions(gadmVersions)
}
\arguments{
\item{gadmVersions}{The GADM versions to validate}
}
\value{
logical vector
}
\description{
Check whether GADM versions are valid
}
\examples{
\dontrun{
Rnightlights:::validGadmVersions("2.8")
}

}
