% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{validNlConfigExtension}
\alias{validNlConfigExtension}
\title{Check if a configName is valid for a given nlType}
\usage{
validNlConfigExtension(extension, configName, nlType)
}
\arguments{
\item{extension}{character the extension of raster}

\item{configName}{character the config shortname of raster}

\item{nlType}{types of nightlight to check}
}
\value{
logical a vector of logical values
}
\description{
Check if a configName is valid for a given nlType
}
\examples{
Rnightlights:::validNlConfigName("VCMCFG", "OLS.Y")
 #returns FALSE

Rnightlights:::validNlConfigName("VCMCFG", "VIIRS.M")
 #returns TRUE

}
