% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{validNlConfigName}
\alias{validNlConfigName}
\title{Check if a configName is valid for a given nlType}
\usage{
validNlConfigName(configName, nlType)
}
\arguments{
\item{configName}{the raster in use}

\item{nlType}{types of nightlight to check}
}
\value{
logical a vector of logical values
}
\description{
Check if a configName is valid for a given nlType
}
\examples{
Rnightlights:::validNlConfigName("VCMCFG", "OLS.Y")
 #returns FALSE
 
Rnightlights:::validNlConfigName("VCMCFG", "VIIRS.M")
 #returns TRUE

}
