% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{validNlPeriod}
\alias{validNlPeriod}
\title{Check if a month number is valid for a given nightlight type}
\usage{
validNlPeriod(nlPeriod, nlType)
}
\arguments{
\item{nlPeriod}{the nlPeriod of interest}

\item{nlType}{type of nightlight either "VIIRS" or "OLS"}
}
\value{
TRUE/FALSE
}
\description{
Check if a month number is valid for a given nightlight type. Note month num is only valid for
"VIIRS" nightlight type
}
\examples{
\dontrun{validNlPeriod("201204","VIIRS")}
 #returns TRUE

\dontrun{validNlPeriod("201203","VIIRS")}
 #returns FALSE

\dontrun{validNlPeriod("2012","OLS")}
 #returns TRUE

}
