% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{validNlPeriods}
\alias{validNlPeriods}
\title{Check if an nlPeriod is valid for a given nightlight type}
\usage{
validNlPeriods(nlPeriods, nlTypes)
}
\arguments{
\item{nlPeriod}{the nlPeriod of interest}

\item{nlType}{type of nightlight}
}
\value{
TRUE/FALSE
}
\description{
Check if an nlPeriod is valid for a given nightlight type
}
\examples{
validNlPeriods("201401","VIIRS.M")
 #returns TRUE

validNlPeriods("201203","VIIRS.M")
 #returns FALSE

validNlPeriods("2012","OLS.Y")
 #returns TRUE

}
