% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{validNlStats}
\alias{validNlStats}
\title{Check if a statistic given is valid}
\usage{
validNlStats(nlStats)
}
\arguments{
\item{nlStat}{the statistic to check}
}
\value{
TRUE/FALSE
}
\description{
Check if a statistic given is valid
}
\examples{

Rnightlights:::validStats("sum")
 #returns TRUE
 
Rnightlights:::validStats("unknownFunction")
 #returns FALSE

}
