% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{validNlTile}
\alias{validNlTile}
\title{Check valid tile number for a given VIIRS nightlight type}
\usage{
validNlTile(nlTile, nlType)
}
\arguments{
\item{nlType}{A character string of nlType}

\item{nlTileNum}{the index of the tile}
}
\value{
TRUE/FALSE
}
\description{
Check if a tile number is valid for a given VIIRS nightlight type.
}
\examples{
Rnightlights:::validNlTile("1", "VIIRS.M")
 #returns TRUE

Rnightlights:::validNlTile("9", "VIIRS.D")
 #returns FALSE

}
