% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{validNlTileNameVIIRS}
\alias{validNlTileNameVIIRS}
\title{Check if a tile index name is valid for a given nightlight type}
\usage{
validNlTileNameVIIRS(tileName)
}
\arguments{
\item{tileName}{the name of the tile}
}
\value{
TRUE/FALSE
}
\description{
Check if a tile number is valid for a given nightlight type. Note tile num is only valid for
"VIIRS" nightlight type
}
\examples{
Rnightlights:::validNlTileNameVIIRS("00N060W")
 #returns TRUE

}
