% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{validNlYearMonthVIIRS}
\alias{validNlYearMonthVIIRS}
\title{Check if a VIIRS nlYearMonth is valid}
\usage{
validNlYearMonthVIIRS(nlYearMonth)
}
\arguments{
\item{nlYearMonth}{the yearmonth in "YYYYMM" format e.g. "201210", "201401"}
}
\value{
TRUE/FALSE
}
\description{
Check if a VIIRS nlYearMonth is valid Note this function is only valid for the "VIIRS" nightlight type
}
\examples{
validNlYearMonthVIIRS("201512")
 returns TRUE

validNlYearMonthVIIRS("201513")
 returns FALSE

validNlYearMonthVIIRS("201201")
 returns FALSE #VIIRS starts in "201204"

}
