% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{validNlYearNum}
\alias{validNlYearNum}
\title{Check if a year is valid for a given nightlight type}
\usage{
validNlYearNum(yearNum, nlType)
}
\arguments{
\item{nlType}{type of nightlight either "VIIRS" or "OLS"}

\item{year}{}
}
\value{
TRUE/FALSE
}
\description{
Check if a year is valid for a given nightlight type
}
\examples{
validNlYearNum("2014","VIIRS")

}

