% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{validStat}
\alias{validStat}
\title{Check if a statistic given is valid}
\usage{
validStat(nlStat)
}
\arguments{
\item{nlStat}{the statistic to check}
}
\value{
TRUE/FALSE
}
\description{
Check if a statistic given is valid
}
\examples{

Rnightlights:::validStat("sum")
 #returns TRUE
 
Rnightlights:::validStat("unknownFunction")
 #returns FALSE

}
