% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{ZonalPipe}
\alias{ZonalPipe}
\title{Create a zonal file if it does not exist and calculate the zonal stats}
\usage{
ZonalPipe(ctryCode, ctryPoly, path.in.shp, path.in.r, path.out.r, path.out.shp,
  zone.attribute, stats)
}
\arguments{
\item{ctryCode}{}

\item{ctryPoly}{}

\item{path.in.shp}{}

\item{path.in.r}{}

\item{path.out.r}{}

\item{path.out.shp}{}

\item{zone.attribute}{}

\item{stat}{}
}
\value{
TRUE/FALSE
}
\description{
Create a zonal file if it does not exist and calculate the zonal stats by calling the 
    myZonal function. Modified from 
    \url{http://www.guru-gis.net/efficient-zonal-statistics-using-r-and-gdal/}
}

