% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{allValidCtryAdmLvls}
\alias{allValidCtryAdmLvls}
\title{Checks if all ctry admLevels are valid}
\usage{
allValidCtryAdmLvls(
  ctryCode = NULL,
  admLevels,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"),
  custPolyPath = NULL
)
}
\arguments{
\item{ctryCode}{\code{character} The ctryCode of the country of interest}

\item{admLevels}{\code{character} The admLevel(s) to search for}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
\code{logical} whether inputted admLevels are valid
}
\description{
Checks if all ctry admLevels are valid
}
\examples{
\dontrun{
validCtryAdmLvls("KEN", "adm0")
#returns "KEN_adm1"
}

}
