% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{createCtryStruct}
\alias{createCtryStruct}
\title{Save ctry admin structure to text file for faster access}
\usage{
createCtryStruct(
  ctryCode = NULL,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"),
  custPolyPath = NULL
)
}
\arguments{
\item{ctryCode}{The ISO3 ctryCode of the country}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to use from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
None
}
\description{
Save ctry admin structure to text file for faster access
}
\examples{
\dontrun{
  Rnightlights:::createCtryStruct("KEN")
}

}
