% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{ctryNameToCode}
\alias{ctryNameToCode}
\title{Convert a country name to its ISO3 code}
\usage{
ctryNameToCode(ctryName)
}
\arguments{
\item{ctryCode}{}
}
\value{
Character full country name
}
\description{
Convert a country name to its ISO3 code. Exposes the rworldmap function rwmGetISO3(ctryName). See the examples
}
\examples{
ctryNameToCode("kenya")
  #returns "KEN"

ctryNameToCode("ken")
  #returns "KEN"

}
