% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrycode.R
\name{ctryNameToCode}
\alias{ctryNameToCode}
\title{Convert country names to ISO3 codes}
\usage{
ctryNameToCode(ctryNames)
}
\arguments{
\item{ctryNames}{Character vector common names of countries to convert}
}
\value{
Character ISO3 ctryCode if found else NA
}
\description{
Convert country names to ISO3 codes. Searches the rworldmap map data.
    With no parameters returns a list of ctryNames and their corresponding
    codes as given by rworldMap
}
\examples{
ctryNameToCode("kenya") #returns "KEN"

ctryNameToCode("ken") #returns "KEN"
  
ctryNameToCode("jamaica") #returns JAM

}
