% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{deleteSavedNlStat}
\alias{deleteSavedNlStat}
\title{Delete a previously saved nlStat}
\usage{
deleteSavedNlStat(nlStatName)
}
\arguments{
\item{nlStatName}{The name of the nlStat to delete}
}
\value{
an nlPeriod vector
}
\description{
Delete a previously saved nlStat
}
\examples{
dateToNlPeriod(dt = "2012-04-01", nlType = "VIIRS.M")
#returns "201204"

}
