% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{dnldCtryPoly}
\alias{dnldCtryPoly}
\title{Download a country's polygon shapefile from \url{http://gadm.org}}
\usage{
dnldCtryPoly(
  ctryCode = NULL,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"),
  custPolyPath = NULL,
  downloadMethod = pkgOptions("downloadMethod")
)
}
\arguments{
\item{ctryCode}{The ISO3 ctryCode of the country polygon to download}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}

\item{downloadMethod}{The method used to download polygons}
}
\value{
TRUE/FALSE Success/Failure of the download
}
\description{
Download a country's polygon shapefile from \url{http://gadm.org}
}
\examples{
\dontrun{
Rnightlights:::dnldCtryPoly("KEN")
}

}
