% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadnltiles.R
\name{downloadNlTilesOLS}
\alias{downloadNlTilesOLS}
\title{Download OLS nightlight tile}
\usage{
downloadNlTilesOLS(nlYear, downloadMethod = pkgOptions("downloadMethod"))
}
\arguments{
\item{nlYear}{the year in "YYYY" format e.g. "2012"}

\item{downloadMethod}{The method to use for download.}
}
\value{
TRUE/FALSE Whether the download was successful
}
\description{
Download OLS nightlight tile
}
\examples{
\dontrun{
if(Rnightlights:::downloadNlTilesOLS("201405"))
  print("download successful")
  }

}
