% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadnltiles.R
\name{downloadNlTilesVIIRS}
\alias{downloadNlTilesVIIRS}
\title{Download VIIRS nightlight tile}
\usage{
downloadNlTilesVIIRS(nlYearMonth, tileNum,
  downloadMethod = pkgOptions("downloadMethod"))
}
\arguments{
\item{nlYearMonth}{the year in "YYYYMM" format e.g. "201401"}

\item{tileNum}{the index of the tile as given by getNlTiles("VIIRS")}

\item{downloadMethod}{The method to use for download.}
}
\value{
TRUE/FALSE Whether the download was successful
}
\description{
Download VIIRS nightlight tile
}
\examples{
\dontrun{
if(downloadNlTilesVIIRS("201401", "1"))
  print("download successful")
  }

}
