% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadnltiles.R
\name{downloadNlTilesVIIRS}
\alias{downloadNlTilesVIIRS}
\title{Download VIIRS nightlight tile}
\usage{
downloadNlTilesVIIRS(
  nlPeriod,
  tileNum,
  downloadMethod = pkgOptions("downloadMethod"),
  nlType,
  configName = pkgOptions(paste0("configName_", nlType))
)
}
\arguments{
\item{nlPeriod}{the nlPeriod of the tile to download}

\item{tileNum}{the index of the tile as given by \code{getNlTiles}}

\item{downloadMethod}{The method to use for download.}

\item{nlType}{A character string of nlType}

\item{configName}{character the type of raster being processed}
}
\value{
TRUE/FALSE Whether the download was successful
}
\description{
Download VIIRS nightlight tile
}
\examples{
\dontrun{
if(Rnightlights:::downloadNlTilesVIIRS("201401", "1"))
  print("download successful")
  }

}
