% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{existsCtryPoly}
\alias{existsCtryPoly}
\title{Checks if a country polygon exists}
\usage{
existsCtryPoly(
  ctryCode = NULL,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"),
  custPolyPath = NULL
)
}
\arguments{
\item{ctryCode}{\code{character} The ctryCode of the country of interest}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\description{
Checks if a country polygon exists
}
\examples{
\dontrun{
existCtryPoly("KEN")
#returns TRUE/FALSE
}

}
