% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{existsPolyFnamePath}
\alias{existsPolyFnamePath}
\title{Check if the decompressed country polygon has been downloaded and stored in the polygon folder}
\usage{
existsPolyFnamePath(ctryCode)
}
\arguments{
\item{ctryCode}{The ctryCode to process}
}
\value{
TRUE/FALSE
}
\description{
Check if the decompressed country polygon has been downloaded and stored in the polygon folder
}
\examples{
Rnightlights:::existsPolyFnamePath("KEN")
 #returns TRUE/FALSE

}
