% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploredata.R
\name{exploreData}
\alias{exploreData}
\title{Run a web application to explore the processed nightlight data cached
    locally}
\usage{
exploreData()
}
\value{
None
}
\description{
Run a web application to perform some exploratory data analysis. The
    application written in shiny either loads demo data or the data 
    processed and cached locally.
}
\examples{
\dontrun{
exploreData()
}

}
