% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{getAllNlConfigNames}
\alias{getAllNlConfigNames}
\title{Generate a list of all possible configNames for a given nlType}
\usage{
getAllNlConfigNames(
  nlTypes = NULL,
  configNames = NULL,
  description = NULL,
  pretty = FALSE
)
}
\arguments{
\item{nlTypes}{if present show only configNames matching the nlTypes}

\item{configNames}{if present show only configNames matching the configNames}

\item{description}{\code{(character)} whether to print the long or
short description. Possible values are:
\itemize{
    \item NULL Do not show descriptions
    \item short Show short descriptions
    \item long Show long descriptions
}}

\item{pretty}{Whether to prettify the output. Prints one line per nlType.
Only available if description is NULL}
}
\description{
Generate a list of all possible configNames for a given nlType
}
\examples{
getAllNlConfigNames("OLS.Y")
 #returns '"cf_cvg", "avg_vis", "stable_lights"'

}
