% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{getAllNlConfigNames}
\alias{getAllNlConfigNames}
\title{Generate a list of all possible configNames for a given nlType}
\usage{
getAllNlConfigNames(nlType)
}
\arguments{
\item{nlType}{if present show only configNames matching the nlType}
}
\description{
Generate a list of all possible configNames for a given nlType
}
\examples{
getAllNlConfigNames("OLS.Y")
 #returns '"cf_cvg", "avg_vis", "stable_lights"'
 
}
