% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAllNlPeriods}
\alias{getAllNlPeriods}
\title{Generate a list of all possible nlPeriods for a given nlType}
\usage{
getAllNlPeriods(nlType)
}
\arguments{
\item{nlType}{type of nightlight either "VIIRS" or "OLS"}
}
\value{
character vector list of nlPeriods
}
\description{
Generate a list of all possible nlPeriods for a given nlType
}
\examples{
\dontrun{getAllNlPeriods("OLS")}
 #returns a vector of all years from 1994 to present

\dontrun{getAllNlPeriods("VIIRS")}
 #returns a vector of all yearMonths from 201204 to present

}
