% Generated by roxygen2: do not edit by hand
% Please edit documentation in shiny_nightlights/nightlights.R
\name{getCtryNlData}
\alias{getCtryNlData}
\title{Returns VIIRS nightlight data for the given ctryCode in the given year months}
\usage{
getCtryNlData(ctryCode, nlYearMonths, stats = pkg_options("stats"), nlType,
  ignoreMissing = NULL, source = "local")
}
\arguments{
\item{ctryCode}{the ISO3 code of the country. Only 1 country can be processed at a time}

\item{nlYearMonths}{a vector of yearMonths}

\item{stats}{a vector of stats. if not supplied defaults to all stats as listed in pkg_options("stats")}

\item{nlType}{the nightlight type i.e. "OLS" or "VIIRS" (default)}

\item{ignoreMissing}{controls how the function behaves if any data is not
found in the data file

\itemize{
    \item NULL (default) only return data if found for all nlYearMonths
       and all stats provided otherwise return NULL
    \item TRUE return any partial data that is found for the provided 
       nlYearMonths and stats. Ignore any missing data
    \item FALSE return all data that is found and call \code{processNtLts}
       to download and process any missing nlYearMonths and stats
}}
}
\value{
dataframe of data for one country in one or multiple yearmonths
}
\description{
Returns VIIRS nightlight data for the given ctryCode and stats in the given 
    year months. Note that getCtryNldata only processes one ctryCode at a time.
    \code{ignoreMissing} plays a significant role here. It can take 3 values:
    
    \itemize{
        \item NULL (default) only return data if found for all nlYearMonths
           and all stats provided otherwise return NULL
        \item TRUE return any partial data that is found for the provided 
           nlYearMonths and stats. Ignore any missing data
        \item FALSE return all data that is found and call \code{processNtLts}
           to download and process any missing nlYearMonths and stats
    }
   
   Farther, if \code{nlYearMonths} is missing, it is assigned values based on
   the value of ignoreMissing. If ignoreMissing is FALSE, nlYearMonths is 
   assigned all existing nlYearMonths to date. This is the equivalent of 
   retrieving all nightlight data for the given country and stats. If 
   ignoreMissing is TRUE or NULL then the existing data is returned.
}
\examples{
#missing stats implies all stats

getCtryNlData("KEN", ignoreMissing=NULL)
 #returns all existing data i.e. all nlYearMonths and all stats for KEN

getCtryNlData("KEN", ignoreMissing=TRUE)
 #same as ignoreMissing=NULL. Returns all existing data i.e. all nlYearMonths
 #and all stats for KEN
 
getCtryNlData("KEN", ignoreMissing=FALSE)
 #for any missing data between 201204 to present download and process the
 #data then return all data
 
getCtryNlData("KEN", nlYearMonths=c("existingNlYearMonth", "missingNlYearMonth"),
    stats=c("existingStat", "missingStat"), ignoreMissing=NULL)
 #Returns NULL
 #(ignoreMissing==NULL returns all data if exists or if any is missing returns NULL)

getCtryNlData("KEN", nlYearMonths=c("existingNlYearMonth", "missingNlYearMonth),
    stats=c("existingStat", "missingStat"), ignoreMissing=TRUE)
 #Returns existingStat for existingNlYearMonth
 #(ignoreMissing==TRUE returns only existing data)
 
getCtryNlData("KEN", nlYearMonths=c("existingNlYearMonth", "missingNlYearMonth),
stats=c("existingStat", "missingStat"), ignoreMissing=FALSE)
 #Runs processNtLts for missingStat in missingNlYearMonth and returns
 #existingStat and missingStat for both existingNlYearMonth and missingNlYearMonth
 #(ignoreMissing==FALSE must return all data: forces processing of any missing)
 
}

