% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrynldata.R
\name{getCtryNlDataColName}
\alias{getCtryNlDataColName}
\title{Construct the name of a nightlight data column given the nightlight type and nlPeriod}
\usage{
getCtryNlDataColName(nlPeriod, nlStat, nlType, configName,
  multiTileStrategy = "all", multiTileMergeFun = "mean",
  removeGasFlares = TRUE)
}
\arguments{
\item{nlPeriod}{character vector The nlPeriod to process}

\item{nlStat}{character vector The stat to be stored in the column}

\item{nlType}{character vector The type of nightlight}

\item{configName}{character the type of raster being processed}

\item{multiTileStrategy}{character How to handle multiple tiles per nlPeriod}

\item{multiTileMergeFun}{character The function to use to merge tiles}

\item{removeGasFlares}{logical Whether to perform gas flare removal pre-processing}
}
\value{
character string
}
\description{
Construct the name of a nightlight data column given the nightlight type and nlPeriod Used in
    creating and retrieving data columns from the nightlight data file
}
\examples{

Rnightlights:::getCtryNlDataColName("201612", "sum", nlType="VIIRS.M")
  
}
