% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getCtryNlDataFnamePath}
\alias{getCtryNlDataFnamePath}
\title{Get the full path to the file containing the country data}
\usage{
getCtryNlDataFnamePath(ctryCode)
}
\arguments{
\item{ctryCode}{}
}
\value{
Character string the full path to the data file of a country
}
\description{
Get the full path to the file containing the country data
}
\examples{
ctryDF <- read.csv(getCtryNlDataFnamePath("KEN"))
 #returns DF with nightlight data for the country

#@export only due to exploreData() shiny app
}
