% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getCtryShpLowestLyrNames}
\alias{getCtryShpLowestLyrNames}
\title{Get the name of the lowest ctry admin level}
\usage{
getCtryShpLowestLyrNames(
  ctryCodes = NULL,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"),
  custPolyPath = NULL
)
}
\arguments{
\item{ctryCodes}{\code{character} ctryCodes the ctryCodes of interest}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
character string The name of the lowest admin level
}
\description{
Get the name of the lowest ctry admin level
}
