% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getCtryStructFnamePath}
\alias{getCtryStructFnamePath}
\title{Construct the full path to the country struct file}
\usage{
getCtryStructFnamePath(
  ctryCode = NULL,
  gadmVersion = pkgOptions("gadmVersion"),
  gadmPolyType = pkgOptions("gadmPolyType"),
  custPolyPath = NULL
)
}
\arguments{
\item{ctryCode}{The ISO3 ctryCode of the country}

\item{gadmVersion}{The GADM version to use}

\item{gadmPolyType}{The format of polygons to download from GADM}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
character string The file path
}
\description{
Construct the full path to the country struct file
}
\examples{
Rnightlights:::getCtryStructFnamePath("KEN")

}
