% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapath.R
\name{getNlDir}
\alias{getNlDir}
\title{Get the paths to the various data locations}
\usage{
getNlDir(dirName)
}
\arguments{
\item{dirName}{character vector The name of the directory to retrieve}
}
\description{
Get the paths to the various locations of nightlights data generated by
    the Rnightlights package. These correspond to the various "dir..." 
    options in the pkgOptions settings
}
\examples{
getNlDir("dirRasterOutput")

getNlDir("dirNlTiles")

getNlDir("dirPolygon")

getNlDir("dirZonals")

}
