% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTifLclNameOLS}
\alias{getNlTifLclNameOLS}
\title{Constructs the filename used to save/access the decompressed OLS .tif file}
\usage{
getNlTifLclNameOLS(nlPeriod)
}
\arguments{
\item{nlPeriod}{the nlPeriod in which the tile was created}
}
\value{
a character vector filename of the .tif VIIRS tile
}
\description{
Constructs the filename used to save/access the decompressed OLS .tif file
}
\examples{
#using default dirNlTiles
\dontrun{
Rnightlights:::getNlTifLclNameOLS("2004")
 #returns "OLS_2004.tif"
 }

}
