% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{getNlTiles}
\alias{getNlTiles}
\title{Create mapping of nightlight tiles}
\usage{
getNlTiles(nlType)
}
\arguments{
\item{nlType}{the nlType of interest}
}
\value{
A data.frame of names of tiles and lon-lat coordinate of top-left corner of each
}
\description{
Creates a data.frame mapping nightlight tile names to their vertice coordinates. This is used to
    identify nightlight tiles as well as to build a spatial polygons dataframe used to plot the tiles. OLS
    only has one tile for the whole world and thus has a dummy entry. OLS is included to
    prevent code duplication by writing separate functions for OLS.
}
\examples{
Rnightlights:::getNlTiles("VIIRS.M")

Rnightlights:::getNlTiles("OLS.Y")

}
