% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTileTifLclNameOLS}
\alias{getNlTileTifLclNameOLS}
\title{Constructs the filename of the decompressed OLS .tif file}
\usage{
getNlTileTifLclNameOLS(
  nlType = "OLS.Y",
  configName = pkgOptions(paste0("configName_", nlType)),
  nlPeriod
)
}
\arguments{
\item{nlType}{character string the nlType}

\item{configName}{character the type of raster being processed}

\item{nlPeriod}{the nlPeriod in which the tile was created}
}
\value{
a character vector filename of the .tif OLS tile
}
\description{
Constructs the filename of the decompressed OLS .tif file
}
\examples{
#using default dirNlTiles
\dontrun{
Rnightlights:::getNlTileTifLclNameOLS("2004")
 #returns "OLS_2004_00N180W.tif"
 }

}
