% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTileTifLclNamePathVIIRS}
\alias{getNlTileTifLclNamePathVIIRS}
\title{Constructs the full path used to save/access the decompressed VIIRS .tif file}
\usage{
getNlTileTifLclNamePathVIIRS(nlPeriod, tileNum, nlType)
}
\arguments{
\item{nlPeriod}{the yearMonth in which the tile was created}

\item{tileNum}{the index of the tile as given in nlTileIndex}

\item{nlType}{The particular VIIRS type e.g. VIIRS_D for daily VIIRS}
}
\value{
a character vector filename of the .tif VIIRS tile
}
\description{
Constructs the full path used to save/access the decompressed VIIRS .tif file
}
\examples{
#using default dirNlTiles
\dontrun{
Rnightlights:::getNlTileTifLclNamePathVIIRS("201401", 1)
 #returns "/dataPath/tiles/VIIRS_2014_01_75N180W.tif"
 }

}
