% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTileTifLclNameVIIRS}
\alias{getNlTileTifLclNameVIIRS}
\title{Constructs the filename of the decompressed VIIRS .tif file}
\usage{
getNlTileTifLclNameVIIRS(nlPeriod, tileNum, nlType)
}
\arguments{
\item{nlPeriod}{the nlPeriod in which the tile was created}

\item{tileNum}{the index of the tile as given in nlTileIndex}

\item{nlType}{character string the nlType}
}
\value{
a character vector filename of the .tif VIIRS tile
}
\description{
Constructs the filename of the decompressed VIIRS .tif file
}
\examples{
#using default dirNlTiles
\dontrun{
Rnightlights:::getNlTileTifLclNameVIIRS("201401", 1, "VIIRS.M")
 #returns "VIIRS_201401_75N180W.tif"
 }

}
