% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlfilename.R
\name{getNlTileZipLclNameOLS}
\alias{getNlTileZipLclNameOLS}
\title{The name with which to save the OLS tile locally}
\usage{
getNlTileZipLclNameOLS(nlPeriod)
}
\arguments{
\item{nlPeriod}{The year of the OLS tile}
}
\value{
character string filename
}
\description{
The name with which to save the OLS tile locally
}
\examples{
\dontrun{
Rnightlights:::getNlTileZipLclNameOLS("2012")
#returns "OLS.Y_2012_00N180W.tar"
}

}
