% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlurl.R
\name{getNlUrlOLS}
\alias{getNlUrlOLS}
\title{Function to return the url of the OLS tile to download}
\usage{
getNlUrlOLS(nlPeriod)
}
\arguments{
\item{nlPeriod}{The nlPeriod of the tile for which to return the tile download URL}
}
\value{
character string Url of the OLS tile file
}
\description{
Function to return the url of the OLS tile to download given the year
}
\examples{
\dontrun{
tileUrl <- Rnightlights:::getNlUrlOLS("1999")
}

}
