% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnlurl.R
\name{getNlUrlVIIRS}
\alias{getNlUrlVIIRS}
\title{Function to return the url of the VIIRS tile to download}
\usage{
getNlUrlVIIRS(
  nlPeriod,
  tileNum,
  nlType,
  configName = pkgOptions(paste0("configName_", nlType))
)
}
\arguments{
\item{nlPeriod}{character string the nlPeriod}

\item{tileNum}{The integer index of the tile to download as given by \code{getNlTiles}}

\item{nlType}{character the nlType to consider}

\item{configName}{character the configName to consider}
}
\value{
Character string Url of the VIIRS tile file
}
\description{
Function to return the url of the VIIRS tile to download given the year, month, and nlTile index
}
\examples{
\dontrun{
tileUrl <- Rnightlights:::getNlUrlVIIRS("20171231", "1", "VIIRS.D")

tileUrl <- Rnightlights:::getNlUrlVIIRS("201401", "1", "VIIRS.M")

tileUrl <- Rnightlights:::getNlUrlVIIRS("2015", "1", "VIIRS.Y")
}

}
