% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{getPolyFname}
\alias{getPolyFname}
\title{Returns the directory name of the unzipped shapefile downloaded from GADM.ORG without the path}
\usage{
getPolyFname(ctryCode)
}
\arguments{
\item{ctryCode}{character the ISO3 code of the country}
}
\value{
character path to zip
}
\description{
Returns the directory name of the unzipped shapefile downloaded from \url{www.GADM.ORG} without the path
}
\examples{
getPolyFnameZip("KEN")
 returns "path/to/"

}

