% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getPolyFname}
\alias{getPolyFname}
\title{Returns the directory name of the unzipped shapefile downloaded from
    http://gadm.org without the path}
\usage{
getPolyFname(ctryCode)
}
\arguments{
\item{ctryCode}{character the ISO3 code of the country}
}
\value{
character name of shapefile directory
}
\description{
Returns the directory name of the unzipped shapefile downloaded from 
    \url{http://gadm.org} without the path
}
\examples{
Rnightlights:::getPolyFname("KEN")
 #returns "KEN_adm_shp"

}
