% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getPolyFnameRDS}
\alias{getPolyFnameRDS}
\title{Get the filename of the polygon zip file as downloaded from \url{http://GADM.ORG}}
\usage{
getPolyFnameRDS(ctryCode = NULL, gadmVersion = pkgOptions("gadmVersion"),
  custPolyPath = NULL)
}
\arguments{
\item{ctryCode}{character the ISO3 code of the country}

\item{gadmVersion}{The GADM version to use}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
character path to zip
}
\description{
Get the filename of the polygon zip file as downloaded from \url{http://GADM.ORG}
}
\examples{
Rnightlights:::getPolyFnameZip("KEN")
 #returns "path/to/"

}
