% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{getPolyFnameZip}
\alias{getPolyFnameZip}
\title{Get the filename of the polygon zip file as downloaded from \url{http://GADM.ORG}}
\usage{
getPolyFnameZip(ctryCode)
}
\arguments{
\item{ctryCode}{character the ISO3 code of the country}
}
\value{
character path to zip
}
\description{
Get the filename of the polygon zip file as downloaded from \url{http://GADM.ORG}
}
\examples{
Rnightlights:::getPolyFnameZip("KEN")
 #returns "path/to/"

}
