% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlCleanup}
\alias{nlCleanup}
\title{Clean up the environment after processing (Not yet implemented)}
\usage{
nlCleanup()
}
\description{
Clean up the environment after processing (Not yet implemented)
}
\examples{
 \dontrun{
 Rnightlights:::nlCleanup()
 }

}
