% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlCleanup}
\alias{nlCleanup}
\title{Clean up the environment after processing (Not yet implemented)}
\usage{
nlCleanup(temp = TRUE, tileCache = FALSE)
}
\arguments{
\item{temp}{boolean should the temp folder be cleared?}

\item{tileCache}{boolean should unnecessary tiles be cleared?}
}
\description{
Clean up the environment after processing (Not yet implemented)
}
\examples{
 \dontrun{
 Rnightlights:::nlCleanup()
 }

}
