% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pkgDefaults}
\alias{pkgDefaults}
\title{Retrieve default global options for the Rnightlights package}
\usage{
pkgDefaults(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]}
pairs to set options. See \code{pkgOptions} for details.}
}
\description{
Retrieve default global options for the Rnightlights package
}
\examples{
#get cropMaskMethod
pkgDefaults("cropMaskMethod") #returns default "rast"

}
