% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pkgOptions}
\alias{pkgOptions}
\title{Set or get options for the Rnightlights package}
\usage{
pkgOptions(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]}
pairs to set options.}
}
\value{
if an option name is supplied as a parameter this returns the 
    value, else a list of all options is returned.
}
\description{
Set or get options for the Rnightlights package
}
\section{Supported options}{

The following options are supported
\describe{
 \item{\code{configName.VIIRS.D}}{(\code{character}) The regex to uniquely
     identify the tile file to use out of the downloaded tile .tgz. The
     version 1 monthly series is run globally using two different 
     configurations.
     
     The first excludes any data impacted by stray light. The second
     includes these data if the radiance vales have undergone the stray-
     light correction procedure (Reference). These two configurations
     are denoted in the filenames as "vcm" and "vcmsl" respectively.
     The "vcmsl" version, that includes the stray-light corrected data,
     will have more data coverage toward the poles, but will be of reduced
     quality.
     
     It is up to the users to determine which set is best for their
     applications. The annual versions are only made with the "vcm"
     version, excluding any data impacted by stray light.}
     
 \item{\code{configName.VIIRS.M}}{(\code{character}) The regex to uniquely
     identify the tile file to use out of the downloaded monthly .tgz
     tile. Has the same options as configName.VIIRS.D}
 
 \item{\code{configName.VIIRS.Y}}{(\code{character}) The regex to uniquely
     identify the tile file to use out of the downloaded tile .tgz. The
     annual products can have other values for the config shortname (Field 5).
     They are:
     \itemize{
       \item vcm-orm \emph{(VIIRS Cloud Mask - Outlier Removed)}: This product
         contains cloud-free average radiance values that have undergone
         an outlier removal process to filter out fires and other ephemeral
         lights.
       \item vcm-orm-ntl \emph{(VIIRS Cloud Mask - Outlier Removed - Nighttime Lights)}:
         This product contains the "vcm-orm" average, with background
         (non-lights) set to zero.
       \item vcm-ntl \emph{(VIIRS Cloud Mask - Nighttime Lights)}: This product
         contains the "vcm" average, with background
         (non-lights) set to zero.}}
 \item{\code{cropMaskMethod}}{(\code{character}) The method to use to 
     clip the nightlight raster tiles to the country boundaries }
 \item{\code{deleteTiles}}{(\code{character}) whether to delete tiles 
     after processing may be useful where diskspace is a concern }
 \item{\code{dirNlData}}{(\code{character}) The directory to store 
     the extracted data files in }
 \item{\code{dirNlRoot}}{\code{character}) The root directory 
     storing the package data}
 \item{\code{dirNlTiles}}{(\code{character}) The directory in which 
     to store the downloaded VIIRS raster tiles }
 \item{\code{dirPolygon}}{(\code{character}) The directory to store 
     the downloaded country administration level polygons }
 \item{\code{dirRasterOutput}}{(\code{character}) The directory in 
     which to store the clipped country rasters }
 \item{\code{dirRasterWeb}}{(\code{character}) The directory in which 
     to store the rasters resampled for web display }
 \item{\code{dirZonals}}{(\code{character}) The directory in which to 
     store the zonal statistics country polygon }
 \item{\code{downloadMethod}}{(\code{character}) The download method 
     to use }
 \item{\code{extractMethod}}{(\code{character}) The method to use to 
     extract data from the rasters }
 \item{\code{gdalCacheMax}}{(\code{character}) The maximum memory gdal 
     should use in gdal_rasterize }
 \item{\code{ntLtsIndexUrlOLS}}{(\code{character}) The url with the OLS 
     tile index }
 \item{\code{ntLtsIndexUrlVIIRS}}{(\code{character}) The url with the 
     VIIRS tile index }
 \item{\code{numCores}}{(\code{character}) The number of processor cores 
     to use when extractMethod = "raster" }
 \item{\code{omitCountries}}{(\code{character}) The countries to exclude 
     in processing }
 \item{\code{stats}}{(\code{character}) The statistics to calculate for 
     country regions. The default are sum and mean. Any other aggregate 
     statistics can be included. Also any aggregate function accessible 
     in the current environment can be added. }
 \item{\code{tmpDir}}{(\code{character}) Change the temporary directory
     for processing rasters. Not in use }
}
}

\examples{
#retrieve the current cropMaskMethod
pkgOptions("cropMaskMethod")

#set the cropMaskMethod
pkgOptions(cropMaskMethod="gdal")

#retrieve all options
pkgOptions()

}
