% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pkgReset}
\alias{pkgReset}
\title{Reset global options for the Rnightlights package}
\usage{
pkgReset()
}
\description{
Reset global options for the Rnightlights package
}
\examples{
#get cropMaskMethod
pkgOptions("cropMaskMethod") #returns default "rast"

#set cropMaskMethod to "gdal"
pkgOptions(cropMaskMethod="gdal") #sets to "gdal"

#check cropMaskMethod has changed
pkgOptions("cropMaskMethod") #returns "gdal"

#reset pkgOptions
pkgReset()

#check cropMaskMethod has been reset
pkgOptions("cropMaskMethod") #returns default "rast"

}
