% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapath.R
\name{removeDataPath}
\alias{removeDataPath}
\title{Deletes a root data path all sub-directories}
\usage{
removeDataPath(dataPath = file.path(getNlDataPath(), ".Rnightlights"))
}
\arguments{
\item{dataPath}{The path to the root folder to be deleted}
}
\value{
None
}
\description{
Deletes a root data path and all sub-directories. It can be the current
    directory or a previously used data path. It will only delete if it
    has the default folder structure of a root data path.
}
\examples{
  \dontrun{
  Rnightlights:::removeDataPath(getNlDataPath())
  }
    
}
