% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{saveNlStat}
\alias{saveNlStat}
\title{Save an nlStat function}
\usage{
saveNlStat(nlStat)
}
\arguments{
\item{nlStat}{character An nlStat}
}
\value{
logical Whether the process was successful
}
\description{
Save an nlStat function to the persistent storage so that it can be
    retrieved and reused later even if the user no longer has its
    definition. Especially useful when saving the results from an
    nlStat so that the function used to retrieve the data is always
    available.
}
\examples{

nlStat <- list("sum", "na.rm=TRUE")

saveNlStat(nlStat)

}
