% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapath.R
\name{setDataVersion}
\alias{setDataVersion}
\title{Add data version file to the root data path}
\usage{
setDataVersion(path = getNlDataPath(),
  pkgVersion = utils::packageDescription("Rnightlights")$Version)
}
\arguments{
\item{path}{The folder to add the README file to}

\item{pkgVersion}{The version of the package}
}
\value{
None
}
\description{
Add data version file to the root data path
}
\examples{
  \dontrun{
  Rnightlights:::setDataVersion(version="0.2.4")
  }
    
}
