% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapath.R
\name{setupDataPath}
\alias{setupDataPath}
\title{Interactively allows the user to set up the default root path}
\usage{
setupDataPath(newDataPath = tempdir(), ...)
}
\arguments{
\item{newDataPath}{Default root path to set}

\item{...}{Not used.}
}
\value{
character string Returns (invisibly) the root path,
    or @NULL if running a non-interactive session.
}
\description{
Interactively allows the user to set up the default root path where data
    is cached
}
\seealso{
Internally, @see "setNlDataPath" is used to set the root path.
    The \code{"base::interactive"} function is used to test whether 
    \code{R} is running interactively or not.
}
