% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{tileName2Idx}
\alias{tileName2Idx}
\title{Get the index of a tile given its name}
\usage{
tileName2Idx(tileName, nlType)
}
\arguments{
\item{tileName}{name as given by getNlTiles()}

\item{nlType}{the nlType of interest}
}
\value{
Integer index of the tile
}
\description{
Get the index of a VIIRS tile as given by getNlTiles() given its name
}
\examples{
Rnightlights:::tileName2Idx("00N060W", "VIIRS.M")

}
