% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrycode.R
\name{validCtryCodes}
\alias{validCtryCodes}
\title{Check if country codes are valid}
\usage{
validCtryCodes(ctryCodes)
}
\arguments{
\item{ctryCodes}{the ISO3 country codes to validate}
}
\value{
named logical vector TRUE/FALSE
}
\description{
Check if country codes are valid
}
\examples{
validCtryCodes(c("KEN", "UGA")) #returns TRUE TRUE

validCtryCodes("UAE") #returns FALSE. "United Arab Emirates" ISO3 code = "ARE"

}
