% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrynldata.R
\name{validCtryNlDataDF}
\alias{validCtryNlDataDF}
\title{Check if a country dataframe is valid}
\usage{
validCtryNlDataDF(ctryNlDataDF)
}
\arguments{
\item{ctryNlDataDF}{the country dataframe}
}
\value{
TRUE/FALSE
}
\description{
Check if a country dataframe is valid
}
\examples{
\dontrun{
  Rnightlights:::validCtryNlDataDF(nlCtryDataDF) #returns TRUE
}
 

}
