% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{validNlPeriod}
\alias{validNlPeriod}
\title{Check if an nlPeriod is valid for a given nightlight type}
\usage{
validNlPeriod(nlPeriod, nlType)
}
\arguments{
\item{nlPeriod}{the nlPeriod of interest}

\item{nlType}{type of nightlight either "VIIRS" or "OLS"}
}
\value{
TRUE/FALSE
}
\description{
Check if an nlPeriod is valid for a given nightlight type
}
\examples{
validNlPeriod("201401","VIIRS")
 #returns TRUE

validNlPeriod("201203","VIIRS")
 #returns FALSE

validNlPeriod("2012","OLS")
 #returns TRUE

}
