% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{validNlPeriodOLS}
\alias{validNlPeriodOLS}
\title{Check if an OLS nlYear is valid}
\usage{
validNlPeriodOLS(nlYear)
}
\arguments{
\item{nlYear}{the year in "YYY" format e.g. "2012"}
}
\value{
TRUE/FALSE
}
\description{
Check if an OLS nlYear is valid
}
\examples{
\dontrun{validNlPeriodOLS("2015")}
 #returns FALSE

\dontrun{validNlPeriodOLS("2004")}
 #returns TRUE

\dontrun{validNlPeriodOLS("201201")}
 #returns FALSE

}
