% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{validNlPeriodVIIRS}
\alias{validNlPeriodVIIRS}
\title{Check if a VIIRS nlYearMonth is valid}
\usage{
validNlPeriodVIIRS(nlYearMonth)
}
\arguments{
\item{nlYearMonth}{the yearmonth in "YYYYMM" format e.g. "201410", "201701"}
}
\value{
TRUE/FALSE
}
\description{
Check if a VIIRS nlYearMonth is valid Note this function is only valid for the "VIIRS" nightlight type
}
\examples{
\dontrun{validNlPeriodVIIRS("201512")}
 #returns TRUE

\dontrun{validNlPeriodVIIRS("201513")}
 #returns FALSE; invalid month 13

\dontrun{validNlPeriodVIIRS("201401")}
 #returns FALSE #VIIRS starts in "201401"

}
