% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{validNlPeriodVIIRS}
\alias{validNlPeriodVIIRS}
\title{Check if a VIIRS nlYearMonth is valid}
\usage{
validNlPeriodVIIRS(nlYearMonth)
}
\arguments{
\item{nlYearMonth}{the yearmonth in "YYYYMM" format e.g. "201210", "201401"}
}
\value{
TRUE/FALSE
}
\description{
Check if a VIIRS nlYearMonth is valid Note this function is only valid for the "VIIRS" nightlight type
}
\examples{
\dontrun{validNlPeriodVIIRS("201512")}
 #returns TRUE

\dontrun{validNlPeriodVIIRS("201513")}
 #returns FALSE

\dontrun{validNlPeriodVIIRS("201201")}
 #returns FALSE #VIIRS starts in "201204"

}
