% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{validNlStats}
\alias{validNlStats}
\title{Check if given statistics are valid}
\usage{
validNlStats(nlStats)
}
\arguments{
\item{nlStats}{the statistics to check}
}
\value{
named logical TRUE/FALSE
}
\description{
Check if given statistics are valid. A valid statistic is one which
   is a function available in the current environment and returns
   a valid value to match.fun
}
\examples{

Rnightlights:::validNlStats(c("sum", "mean"))
 #returns TRUE TRUE
 
Rnightlights:::validNlStats("unknownFunction")
 #returns FALSE

}
