% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{validNlTileNameVIIRS}
\alias{validNlTileNameVIIRS}
\title{Check if a tile index name is valid for a given nightlight type}
\usage{
validNlTileNameVIIRS(tileName)
}
\arguments{
\item{nlTileNum}{the index of the tile}

\item{nlType}{type of nightlight either "VIIRS" or "OLS"}
}
\value{
TRUE/FALSE
}
\description{
Check if a tile number is valid for a given nightlight type. Note tile num is only valid for
"VIIRS" nightlight type
}
\examples{
validNlTileNameVIIRS("00N060W","VIIRS")
 returns TRUE

validNlTileNameVIIRS("9","VIIRS")
 returns FALSE

validNlMonthName("","OLS")
 returns FALSE

}

