% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{validNlTileNameVIIRS}
\alias{validNlTileNameVIIRS}
\title{Check valid VIIRS nightlight tile name}
\usage{
validNlTileNameVIIRS(tileName, nlType)
}
\arguments{
\item{tileName}{the name of the tile}

\item{nlType}{character the nlType}
}
\value{
TRUE/FALSE
}
\description{
Check if a tile name is valid for a given VIIRS nightlight type.
}
\examples{
Rnightlights:::validNlTileNameVIIRS("00N060W", "VIIRS.M")
 #returns TRUE

}
