% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{validNlTileNumVIIRS}
\alias{validNlTileNumVIIRS}
\title{Check if a tile index number is valid for a given nightlight type}
\usage{
validNlTileNumVIIRS(nlTileNum)
}
\arguments{
\item{nlTileNum}{the index of the tile}
}
\value{
TRUE/FALSE
}
\description{
Check if a tile number is valid for a given nightlight type. Note tile num is only valid for
"VIIRS" nightlight type
}
\examples{
Rnightlights:::validNlTileNumVIIRS("1")
 #returns TRUE

Rnightlights:::validNlTileNumVIIRS("9")
 #returns FALSE

}
