% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{validNlTileNumVIIRS}
\alias{validNlTileNumVIIRS}
\title{Check valid tile number for a given VIIRS nightlight type}
\usage{
validNlTileNumVIIRS(nlTileNum, nlType)
}
\arguments{
\item{nlTileNum}{the index of the tile}

\item{nlType}{A character string of nlType}
}
\value{
TRUE/FALSE
}
\description{
Check if a tile number is valid for a given VIIRS nightlight type.
}
\examples{
Rnightlights:::validNlTileNumVIIRS("1", "VIIRS.M")
 #returns TRUE

Rnightlights:::validNlTileNumVIIRS("9", "VIIRS.D")
 #returns FALSE

}
