% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nltype.R
\name{validNlTypes}
\alias{validNlTypes}
\title{Checks if a given character string is a valid nlType}
\usage{
validNlTypes(nlTypes)
}
\arguments{
\item{nlTypes}{\code{character vector} The nlTypes}
}
\value{
\code{logical} whether the strings are valid nlTypes
}
\description{
Checks if a given character string is a valid nlType
}
\examples{
validNlTypes("VIIRS.D") #returns TRUE

validNlTypes("VIIRS") #returns FALSE

}
