% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlperiod.R
\name{validNlYearNum}
\alias{validNlYearNum}
\title{Check if a year is valid for a given nightlight type}
\usage{
validNlYearNum(yearNum, nlType)
}
\arguments{
\item{yearNum}{the year of interest}

\item{nlType}{type of nightlight either "VIIRS" or "OLS"}
}
\value{
TRUE/FALSE
}
\description{
Check if a year is valid for a given nightlight type
}
\examples{
\dontrun{validNlYearNum("2014","VIIRS")}

}
