% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightlights.R
\name{validStat}
\alias{validStat}
\title{Check if a statistic given is valid}
\usage{
validStat(stat)
}
\arguments{
\item{stat}{the statistic to check}
}
\value{
TRUE/FALSE
}
\description{
Check if a statistic given is valid
}
\examples{

validStat("sum")
 returns TRUE

}
