% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ZonalPipe}
\alias{ZonalPipe}
\title{Create a zonal file if it does not exist and calculate the zonal stats}
\usage{
ZonalPipe(ctryCode, admLevel, ctryPoly, path.in.shp, path.in.r, path.out.r,
  path.out.shp, zone.attribute, nlStats,
  gadmVersion = pkgOptions("gadmVersion"), custPolyPath = NULL)
}
\arguments{
\item{ctryCode}{the ctryCode of interest}

\item{admLevel}{The country admin level of interest}

\item{ctryPoly}{the SpatialPolygonsDataFrame country polygon to process}

\item{path.in.shp}{The path to the country shapefile}

\item{path.in.r}{The path to the raster tile}

\item{path.out.r}{The path where to save the output zonal raster}

\item{path.out.shp}{The path to save the output zonal shapefile (Ignored)}

\item{zone.attribute}{The zonal attribute to calculate}

\item{nlStats}{The stats to calculate}

\item{gadmVersion}{The GADM version to use}

\item{custPolyPath}{Alternative to GADM. A path to a custom shapefile zip}
}
\value{
TRUE/FALSE
}
\description{
Create a zonal file if it does not exist and calculate the zonal stats by calling the 
    myZonal function. Modified from 
    \url{http://www.guru-gis.net/efficient-zonal-statistics-using-r-and-gdal/}
}
